### 2023-10-30 #147 @nabobalis
`sunpy` Review
Submitting Author: Nabil Freij (@nabobalis)
All current maintainers: @ehsteve,@dpshelio,@wafels,@ayshih,@Cadair,@nabobalis,@dstansby,@DanRyanIrish,@wtbarnes,@ConorMacBride,@alasdairwilson,@hayesla,@vn-ki
Package Name: sunpy
One-Line Description of Package: Python for Solar Physics
Repository Link: https://github.com/sunpy/sunpy
Version submitted: 5.0.1
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8384174.svg)](https://doi.org/10.5281/zenodo.8384174)
Version accepted: 5.1.1

---

- sunpy is a community-developed, free and open-source solar data analysis environment for Python. It includes an interface for searching and downloading data from multiple data providers, data containers for image and time series data, commonly used solar coordinate frames and associated transformations, as well as other functionality needed for solar data analysis.

  **I will need to double check the examples, we have documentation for all public API**

## Comments from Nabil

Hopefully everything is correct!
I did prune the JOSS and editions parts of the template, I hope that is ok since they are not relevant for this review.
