### 2023-10-30 bea3c9d
[pre-commit.ci] pre-commit autoupdate (#7276)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
