### 2023-10-31 2e195bd
Add whatsnew entry
