### 2023-10-31 1248b42
Add whatsnew entry
