### 2023-10-31 eb2b9c7
Update mapbase.py - Add 'fill' keyword to draw_contours function.
Added the 'fill' keyword to the draw_contours function, enabling the option to display filled contours using `~matplotlib.axes.Axes.contourf` in addition to the standard line contours. This enhancement provides users more flexibility in visualizing contour plots.
