### 2023-11-01 1bc4ce5
Bump stefanzweifel/git-auto-commit-action from 4 to 5 (#7283)
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
