+++ .github/workflows/changelog_bot.yml
@@ -31,7 +31,7 @@ jobs:
           PR_BODY: ${{ github.event.pull_request.body }}
           BOT_AUTHOR: ${{ env.BOT_USER_NAME }} <${{ env.BOT_USER_EMAIL }}>
       - name: Commit changelog
-        uses: stefanzweifel/git-auto-commit-action@v4
+        uses: stefanzweifel/git-auto-commit-action@v5
         with:
           commit_message: '[skip ci] Add changelog'
           add_options: '-A'
