### 2023-11-01 488e74f
Fixed a copy-and-paste error in the ASDF manifest (#7282)
