### 2023-11-01 4579079
Backport PR #7282 on branch 5.0 (Fixed a copy-and-paste error in the ASDF manifest) (#7284)
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
