+++ changelog/7282.trivial.rst
@@ -1,0 +1,1 @@
+Fixed our ASDF manifest having the incorrect ID.
