### 2023-11-03 #147 @isabelizimm
Thank you kindly for this submission, @nabobalis 🙌  I also appreciate your preparedness beforehand going through our checklists!

---
---

The link [@sunpy/sunpy-developers](https://github.com/orgs/sunpy/teams/sunpy-maintainers) shows as a 404 for me. My guess is it's a view that needs permissions. Is there an alternative list that is public you can link to?

Your editor for this review will be @cmarmo
