### 2023-11-03 #147 @nabobalis
Thank you @isabelizimm for pre-reviewing the submission.

> ## Editor comments
> The link [@sunpy/sunpy-developers](https://github.com/orgs/sunpy/teams/sunpy-maintainers) shows as a 404 for me. My guess is it's a view that needs permissions. Is there an alternative list that is public you can link to?

I will list the names out as the team has not changed in a while, I am surprised that it isn't public. I will see if I can get that fixed.
