### 2023-11-06 a0d807c
Update .pre-commit-config.yaml
