+++ .pre-commit-config.yaml
@@ -1,5 +1,4 @@
 exclude: "\\.asdf$"
-
 repos:
   # This should be before any formatting hooks like isort
   - repo: https://github.com/astral-sh/ruff-pre-commit
@@ -7,14 +6,12 @@ repos:
     hooks:
       - id: ruff
         args: ["--fix"]
-
   - repo: https://github.com/PyCQA/isort
     rev: 5.12.0
     hooks:
       - id: isort
         args: ["--sp", "setup.cfg"]
         exclude: ".*(.fits|.fts|.fit|.header|.txt|tca.*|extern.*|.rst|.md|cm/__init__.py|sunpy/extern|docs/conf.py)$"
-
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.5.0
     hooks:
@@ -30,13 +27,11 @@ repos:
         exclude: ".*(.fits|.fts|.fit|.header|.txt|tca.*|.json)$|^CITATION.rst$"
       - id: mixed-line-ending
         exclude: ".*(.fits|.fts|.fit|.header|.txt|tca.*)$"
-
   - repo: https://github.com/codespell-project/codespell
     rev: v2.2.6
     hooks:
       - id: codespell
         args: ['--config setup.cfg --write-changes']
-
   - repo: local
     hooks:
       - id: generate-sunpy-net-hek-attrs
@@ -51,7 +46,6 @@ repos:
         language: system
         pass_filenames: false
         always_run: true
-
-
 ci:
   autofix_prs: false
+  autoupdate_schedule: "quarterly"
