### 2023-11-08 767b302
Backport PR #7247: Edit sunpy.io._cdf.read_cdf to avoid Pandas PerformanceWarning: DataFrame is highly fragmented (fix #7246)
