### 2023-11-08 d6dc501
Improve handling of WISPR maps (#7180)
* Supply nice detector names for WISPR maps

* Supply improved colormap defaults for WISPR maps

* Make 'level' handling more robust for WISPR maps

* Add changelog

* Address review comments
