### 2023-11-08 12b62fa
Add unit tests for frames, including triggering downloads
