### 2023-11-08 4cc4a69
Update docs/whatsnew/5.1.rst
Co-authored-by: David Stansby <dstansby@gmail.com>
