### 2023-11-08 60f0fa5
Update docs/whatsnew/5.1.rst
Co-authored-by: David Stansby <dstansby@gmail.com>
