### 2023-11-08 1bb57f2
Update docs/tutorial/installation.rst
Co-authored-by: Will Barnes <will.t.barnes@gmail.com>
