### 2023-11-08 e4f5efe
Add missing "What's New" entries (#7288)
Co-authored-by: David Stansby <dstansby@gmail.com>
