### 2023-11-08 65dcf90
Register both the old manifest and the new manifest
