### 2023-11-10 b91cb22
Update precommit and make sure docs/conf is ok
