### 2023-11-10 09142e1
Release statistics  (#7290)
* Add release statistics to "What's New"

* Update precommit and make sure docs/conf is ok
