+++ changelog/7206.bugfix.rst
@@ -1,1 +1,0 @@
-Fixed a bug with :func:`~sunpy.map.sample_at_coords()` where sampling outside the bounds of the map would sometimes not error and instead return strange pixel values.
