+++ changelog/7143.feature.rst
@@ -1,1 +1,0 @@
-Add a keyword (``map_center_longitude``) to :func:`~sunpy.map.header_helper.make_heliographic_header` for centering the heliographic map at a longitude other than zero longitude.
