+++ changelog/7202.removal.rst
@@ -1,2 +1,0 @@
-``sunpy.visualisation.limb.draw_limb()`` has been removed.
-Use :func:`sunpy.visualization.drawing.limb` instead.
