### 2023-11-10 e5ed12d
fix asdf schema examples
