### 2023-11-15 39ae1b7
Backport PR #7292: fix asdf schema examples
