### 2023-11-15 b7c2d25
fix asdf schema examples (#7292)
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
