+++ changelog/7292.trivial.rst
@@ -1,0 +1,1 @@
+Fix example formatting in a few asdf schemas.
