### 2023-11-17 22fe316
add _priority attribute to handle multiple possible map sources
