### 2023-11-17 7c22a18
add _priority attribute to handle multiple possible map sources
