### 2023-11-17 7b54438
add _priority attribute to handle multiple possible map sources
