### 2023-11-17 8774c1c
fix changelog formatting
