+++ changelog/7299.feature.rst
@@ -1,2 +1,2 @@
-Added a `_priority` attribute to `~sunpy.map.GenericMap` to provide a way to decide when more than one
+Added a ``_priority`` attribute to `~sunpy.map.GenericMap` to provide a way to decide when more than one
 source is found for a given map input.
