### 2023-11-17 fdaa3a7
Update sunpy/map/map_factory.py
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
