### 2023-11-17 33712f3
Do not add pfsspy GONG map source to registry to avoid multiple-match error
