### 2023-11-17 563308c
more explicit check on class name and module
