### 2023-11-17 8767c04
Pre-commit + CI + devdeps (sunpy/sunpy-soar#111)
* Pre-commit

* Tweak test and add devdeps

* Changelog for 1.10

* Add changelog readme
