### 2023-11-17 56fea30
Add note about map class to whatsnew
