### 2023-11-17 b39b72c
Backport PR #7302: Tweak to example because of order of operations
