### 2023-11-17 e959004
Backport PR #7300 on branch 5.1 (Do not add `pfsspy` GONG map source to registry to avoid `MultipleMatchError`) (#7306)
Co-authored-by: David Stansby <dstansby@gmail.com>
