### 2023-11-18 5ab1052
Backport PR #7308: Pin drms to < 0.7
