### 2023-11-18 755d53f
Pin drms to < 0.7 (#7308)
* Pin drms to < 0.7

* Add changelog
