+++ changelog/7307.trivial.rst
@@ -1,0 +1,3 @@
+Progress from :meth:`sunpy.net.jsoc.JSOCClient.fetch` is now reported through the logger.
+Pinned the ``drms`` requirement to ``>= 0.7`` to account for this breaking API change in ``drms``.
+Note that this change supersedes that in `PR #7308 <https://github.com/sunpy/sunpy/pull/7308>`_.
