### 2023-11-20 df94c5e
Merge pull request #7311 from dstansby/pin-drms
Use v0.6 of drms for intershpinx

(cherry picked from commit ead8119397e2fc8ed8e0ebdab587b682d995c679)
