+++ changelog/7068.trivial.rst
@@ -1,1 +1,0 @@
-Add a Dependabot config file to auto-update GitHub action versions.
