+++ changelog/7250.removal.rst
@@ -1,2 +1,0 @@
-The deprecated ``sunpy.data.download_sample_data()`` has been removed
-Use :func:`sunpy.data.sample.download_all` instead.
