### 2023-11-22 fc4311f
added parenthesis to plot line example
