### 2023-11-22 293081b
Refactor get_horizons_coord() to not use astroquery
