+++ sunpy/coordinates/ephemeris.py
@@ -1,8 +1,11 @@
 """
 Ephemeris calculations using SunPy coordinate frames
 """
+import re
+import tempfile
+
 import numpy as np
-from packaging import version
+import requests
 
 import astropy.units as u
 from astropy.constants import c as speed_of_light
@@ -18,6 +21,7 @@ from astropy.coordinates.representation import (
     CartesianRepresentation,
     SphericalRepresentation,
 )
+from astropy.io import ascii
 from astropy.time import Time
 
 from sunpy import log
@@ -198,18 +202,19 @@ def get_horizons_coord(body, time='now', id_type=None, *, include_velocity=False
     solar-system body at a specified time.  This location is the instantaneous or "true" location,
     and is not corrected for light travel time or observer motion.
 
-    .. note::
-        This function requires the Astroquery package to be installed and
-        requires an Internet connection.
-
     Parameters
     ----------
     body : `str`
         The solar-system body for which to calculate positions.  One can also use the search form
         linked below to find valid names or ID numbers.
     id_type : `None`, `str`
-        See the astroquery documentation for information on id_types: `astroquery.jplhorizons`.
-        If the installed astroquery version is less than 0.4.4, defaults to ``'majorbody'``.
+        Defaults to `None`, which searches major bodies first, and then searches
+        small bodies (comets and asteroids) if no major body is found.  If
+        ``'smallbody'``, the search is limited to only small bodies.  If
+        ``'designation'``, the search is limited to only small-body designations.
+        If ``'name'``, the search is limited to only small-body names.  If
+        ``'asteroid_name'`` or ``'comet_name'``, the search is limited to only
+        asteroid names or only comet names, respectively.
     time : {parse_time_types}, `dict`
         Time to use in a parse_time-compatible format.
 
@@ -219,8 +224,7 @@ def get_horizons_coord(body, time='now', id_type=None, *, include_velocity=False
         ``start_time`` and ``stop_time`` must be in a parse_time-compatible format,
         and are interpreted as UTC time. ``step`` must be a string with either a
         number and interval length (e.g. for every 10 seconds, ``'10s'``), or a
-        plain number for a number of evenly spaced intervals. For more information
-        see the docstring of `astroquery.jplhorizons.HorizonsClass`.
+        plain number for a number of evenly spaced intervals.
 
     include_velocity : `bool`, optional
         If True, include the body's velocity in the output coordinate.  Defaults to False.
@@ -240,11 +244,12 @@ def get_horizons_coord(body, time='now', id_type=None, *, include_velocity=False
     ----------
     * `JPL HORIZONS <https://ssd.jpl.nasa.gov/?horizons>`_
     * `JPL HORIZONS form to search bodies <https://ssd.jpl.nasa.gov/horizons.cgi?s_target=1#top>`_
-    * `Astroquery <https://astroquery.readthedocs.io/en/latest/>`_
 
     Examples
     --------
-    >>> from sunpy.coordinates.ephemeris import get_horizons_coord
+    .. minigallery:: sunpy.coordinates.get_horizons_coord
+
+    >>> from sunpy.coordinates import get_horizons_coord
 
     Query the location of Venus
 
@@ -256,7 +261,7 @@ def get_horizons_coord(body, time='now', id_type=None, *, include_velocity=False
     Query the location of the SDO spacecraft
 
     >>> get_horizons_coord('SDO', '2011-11-11 11:11:11')  # doctest: +REMOTE_DATA
[…]
