### 2023-11-22 #147 @cmarmo
Hi @nabobalis,

thank you for your patience!

Let's welcome here @Septaris and @nutjob4life who kindly volunteered to review SunPy for pyOpenSci! :wave:
Thank you both! Please take some time to introduce yourself here before starting the review.

- [ ] Sonny Lion survey completed.
- [ ] Sean Kelly survey completed.

2. Here is the **[review template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template)** that you will need to fill out and submit here as a comment, once your review is complete.

I'm going to ask reviewers to complete reviews in three weeks, the review for SunPy is then due for **December 13th**, please let me know if you have any major issue preventing to meet this deadline.

Also, please get in touch with any questions or concerns!

Thanks again for your commitment! :pray:
