### 2023-11-22 caab861
Remove astroquery as a library dependency
