+++ sunpy-dev-env.yml
@@ -28,7 +28,6 @@ dependencies:
   - zeep
 
   # Optional
-  - astroquery
   - dask
   - glymur
   - jplephem
@@ -53,6 +52,7 @@ dependencies:
   - tox-conda
 
   # Documentation
+  - astroquery
   - graphviz
   - ruamel.yaml
   - sphinx
