+++ tox.ini
@@ -68,7 +68,6 @@ deps =
     oldestdeps: tqdm<4.33.0
     oldestdeps: zeep<3.5.0
     # These are specific extras we use to run in specific factors
-    online: astroquery
     online: pytest-rerunfailures
     online: pytest-timeout
     online: matplotlib>=3.8
