### 2023-11-23 e73dfde
Remove sunpy.database
