+++ sunpy/database/tests/test_tables.py
@@ -1,577 +1,0 @@
-# Author: Simon Liedtke <liedtke.simon@googlemail.com>
-#
-# This module was developed with funding provided by
-# the Google Summer of Code (2013).
-
-import os
-from datetime import datetime
-from collections.abc import Hashable
-
-import numpy as np
-import pytest
-
-import astropy.units as u
-from astropy import conf
-
-from sunpy.data.test import get_test_filepath
-from sunpy.data.test import rootdir as testdir
-from sunpy.database import Database
-from sunpy.database.tables import (
-    DatabaseEntry,
-    FitsHeaderEntry,
-    FitsKeyComment,
-    Tag,
-    WaveunitNotFoundError,
-    _create_display_table,
-    entries_from_dir,
-    entries_from_fido_search_result,
-    entries_from_file,
-    entries_from_query_result,
-)
-from sunpy.net import Fido
-from sunpy.net import attrs as net_attrs
-from sunpy.net import vso
-
-RHESSI_IMAGE = get_test_filepath('hsi_image_20101016_191218.fits')
-GOES_DATA = get_test_filepath('go1520110607.fits')
-
-
-@pytest.fixture
-def mq_image(waveunit_fits_directory):
-    fits_files = waveunit_fits_directory.glob('*.fits')
-    return os.fspath([f for f in fits_files if f.name == 'mq130812.084253.fits'][0])
-
-
-@pytest.fixture
-def eit_195_image(eit_fits_directory):
-    fits_files = eit_fits_directory.glob('*.fits')
-    return os.fspath([f for f in fits_files if f.name == 'efz20040301.000010_s.fits'][0])
-
-
-@pytest.fixture
-def fido_search_result():
-    # A search query with responses from all instruments
-    # No JSOC query
-    return Fido.search(
-        net_attrs.Time("2012/1/1", "2012/1/2"),
-        net_attrs.Instrument('lyra') & net_attrs.Level.two | net_attrs.Instrument('eve') |
-        net_attrs.Instrument('XRS') | net_attrs.Instrument('noaa-indices') |
-        net_attrs.Instrument('noaa-predict') |
-        (net_attrs.Instrument('norh') & net_attrs.Wavelength(17 * u.GHz)) |
-        (net_attrs.Instrument('rhessi') & net_attrs.Physobs("summary_lightcurve"))
-    )
-
-
-@pytest.fixture
-def query_result():
-    client = vso.VSOClient()
-    return client.search(net_attrs.Time('2001/1/1', '2001/1/2'),
-                         net_attrs.Instrument('EIT'),
-                         response_format="legacy")
-
-
-@pytest.fixture
-def qr_with_none_waves():
-    return vso.VSOClient().search(
-        net_attrs.Time('20121224T120049.8', '20121224T120049.8'),
-        net_attrs.Provider('SDAC'), net_attrs.Instrument('VIRGO'),
-        response_format="legacy")
-
-
-@pytest.fixture
-def qr_block_with_missing_physobs():
-    return vso.VSOClient().search(
-        net_attrs.Time('20130805T120000', '20130805T121000'),
-        net_attrs.Instrument('SWAVES'), net_attrs.Source('STEREO_A'),
-        net_attrs.Provider('SSC'), net_attrs.Wavelength(
-            10 * u.kHz, 160 * u.kHz),
-        response_format="legacy")[0]
-
-
-@pytest.fixture
-def qr_block_with_kev_unit():
-    return vso.VSOClient().search(
[…]
