+++ setup.cfg
@@ -53,8 +53,6 @@ asdf =
   asdf>=2.8.0
 dask =
   dask[array]>=2021.4.0
-database =
-  sqlalchemy>=1.3.4
 image =
   scikit-image>=0.18.0
   # !=1.10.0 due to https://github.com/scipy/scipy/issues/17718
@@ -90,7 +88,6 @@ visualization =
 all =
   %(asdf)s
   %(dask)s
-  %(database)s
   %(image)s
   %(jpeg2000)s
   %(map)s
@@ -175,7 +172,6 @@ filterwarnings =
     error
     # Do not fail on pytest config issues (i.e. missing plugins) but do show them
     always::pytest.PytestConfigWarning
-    ignore:The sunpy.database module is no longer actively maintained
     # A list of warnings to ignore follows. If you add to this list, you MUST
     # add a comment or ideally a link to an issue that explains why the warning
     # is being ignored
@@ -229,8 +225,6 @@ filterwarnings =
     ignore:'cgi' is deprecated and slated for removal in Python 3.13:DeprecationWarning:zeep.utils
     # Deprecation warning added in NumPy 1.24, triggered by scikit-image
     ignore:.*np.bool8.*is a deprecated alias.*:DeprecationWarning
-    # Warnings for sqlalchemy 2.0. For now we pin the version and ignore this warning
-    ignore:Deprecated API features detected.*SQLAlchemy 2
     # See https://github.com/astropy/astropy/issues/14509 and
     # https://github.com/h5py/h5py/issues/2243 for upstream issues about this warning being
     # emitted
