### 2023-11-27 2673e5f
Add requests as an explicit core dependency
