+++ sunpy-dev-env.yml
@@ -20,6 +20,7 @@ dependencies:
   - python_abi
   - python-dateutil
   - reproject
+  - requests
   - scikit-image
   - scipy
   - setuptools
