### 2023-11-27 b16dfc0
Add changelog entries
