+++ changelog/7319.bugfix.rst
@@ -1,0 +1,1 @@
+Long object names are no longer truncated in the logging output of :func:`~sunpy.coordinates.get_horizons_coord`.
