+++ .github/workflows/scheduled_builds.yml
@@ -16,6 +16,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - run: gh workflow run ci.yml --repo sunpy/sunpy --ref main
+      - run: gh workflow run ci.yml --repo sunpy/sunpy --ref 5.1
       - run: gh workflow run ci.yml --repo sunpy/sunpy --ref 5.0
     env:
       GITHUB_TOKEN: ${{ secrets.WORKFLOW_TOKEN }}
