### 2023-11-27 a9cedc3
Backport PR #7316 on branch 5.0 (Fixed the superpixel sample) (#7321)
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
