+++ examples/map/map_resampling_and_superpixels.py
@@ -13,15 +13,16 @@ import astropy.units as u
 import sunpy.data.sample
 import sunpy.map
 
-###############################################################################
+##############################################################################
 # We start with the sample data.
 
 aia_map = sunpy.map.Map(sunpy.data.sample.AIA_171_IMAGE)
 
 ##############################################################################
-# To reduce the angular resolution of the map you can use the `~sunpy.map.GenericMap.resample` method,
-# specifying the new dimensions in pixels. By default, this method uses linear interpolation
-# but this can be changed with the ``method`` argument ('nearest', 'linear' or 'spline').
+# To reduce the angular resolution of the map, you can use the
+# :meth:`~sunpy.map.GenericMap.resample` method, specifying the new dimensions
+# in pixels. By default, this method uses linear interpolation but this can be
+# changed with the ``method`` argument ('nearest', 'linear' or 'spline').
 
 new_dimensions = [40, 40] * u.pixel
 aia_resampled_map = aia_map.resample(new_dimensions)
@@ -35,14 +36,15 @@ aia_resampled_map.plot(axes=ax)
 plt.show()
 
 ##############################################################################
-# Another way to resample is by using the `~sunpy.map.GenericMap.superpixel` method.
-# This can be used to increase the signal to noise ratio by reducing the
-# resolution of the image by combining pixels. This means that the new dimension
-# must divide the original size exactly.
-# For example you can reduce the AIA map resolution by a factor of 16.
+# Another way to reduce the angular resolution of the map is by using the
+# :meth:`~sunpy.map.GenericMap.superpixel` method, which combines pixels.
+# The superpixel dimensions do not need to be square, and the intensity of
+# each superpixel defaults to the sum of the constituent pixels. For example,
+# you can reduce the AIA map resolution by a factor of 16 by specifying 16x16
+# superpixels.
 
-new_dimensions = u.Quantity(aia_map.dimensions) / 16
-aia_superpixel_map = aia_map.superpixel(new_dimensions)
+superpixel_size = [16, 16] * u.pixel
+aia_superpixel_map = aia_map.superpixel(superpixel_size)
 
 ##############################################################################
 # Let's plot the result.
