### 2023-11-27 #147 @nutjob4life
Hi folks; it was the Thanksgiving holiday here in the United States and I'm back to work today. Naturally I've got 497 work emails to get through but I'm looking forward to helping out.

Since @Septaris provided some background I will too: I've been with [NASA](https://www.nasa.gov/)'s [Jet Propulsion Laboratory](https://www.jpl.nasa.gov/) for over 20 years now and I work with the [Planetary Data System](https://pds.nasa.gov/) developing mostly Python-based software for handling large-scale atmospheric, geophysical, planetary/plasma, small bodies, and other scientific data.

I'm also a contributor to numerous open source Python projects outside of science, and Python packaging is a passion, though so I'll try to be extra thorough in that regard.
