### 2023-11-28 #147 @nutjob4life
Estimated hours spent reviewing: 1.25 hours

---

First off, bravo. This is an incredible package that's well-documented and well-assembled. And it's feature-packed! Nice! It's an honor to be able to review it.

My notes follow:

1. The "vignette" on the package's homepage `README.md` works only if you're using Conda. It should work with a bare Python:
$ python3 -m venv sunpy
$ sunpy/bin/pip install --quiet --upgrade pip
$ sunpy/bin/pip install --quiet sunpy==5.0.1
$ sunpy/bin/python
Python 3.11.6 (main, Oct  2 2023, 13:45:54) [Clang 15.0.0 (clang-1500.0.40.1)] on darwin
>>> import sunpy.map
  File "/private/tmp/sunpy/lib/python3.11/site-packages/sunpy/map/__init__.py", line 7, in <module>
    from sunpy.map.mapbase import *
  File "/private/tmp/sunpy/lib/python3.11/site-packages/sunpy/map/mapbase.py", line 15, in <module>
ModuleNotFoundError: No module named 'matplotlib'
```
Maybe the `README.md` could mention that the `[map]` extra is required for the vignette? This also applies to the "Any additional setup required to use the package" criterion.

2. The API is _enormous_, so I can only spot-check if there's "Function Documentation: for all user-facing functions".

3. The same applies to "Examples for all user-facing functions", however I find that `sunpy.util.expand_list` does not have an example. (This also applies to the "All functions have documentation and associated examples for use" criterion.)

4. Python packaging has been (sadly) a moving target, but this project seems to find the balance between `setup.cfg` and `pyproject.toml`. (The Python Packaging Authority has finally settled on moving away from `setup.py` and `setup.cfg` so perhaps later versions of `sunpy` can move all project metadata to `pyproject.toml` in a future version.)

5. "Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file". There is indeed one vignette in the `README.md` however I would _not_ call this a small package! 😆

6. "Any performance claims of the software been confirmed" … the documentation doesn't make any mention of performance, so this criterion easily passes 😇.

7. Running `flake8 --count sunpy` for "Code format is standard throughout package and follows PEP 8 guidelines" shows 501 violations

8. JOSS criteria: version 1.0.8 was reviewed by JOSS, but this version (5.0.1) does not have `paper.md`. Is version 5.0.1 going to be submitted to JOSS? If so, we'll need this file. If not, then no worries 😉

That's all I got! Thanks for the opportunity to review this.
