### 2023-11-28 #147 @nabobalis
Thank you @nutjob4life for reviewing the package!

If I am allowed to respond to some of the comments, I will do so now, if not, I can delete this after.

> #### Review Comments
> First off, bravo. This is an incredible package that's well-documented and well-assembled. And it's feature-packed! Nice! It's an honor to be able to review it.
>
> My notes follow:
>
> 1. The "vignette" on the package's homepage `README.md` works only if you're using Conda. It should work with a bare Python:
>
> ```
> $ python3 -m venv sunpy
> $ sunpy/bin/pip install --quiet --upgrade pip
> $ sunpy/bin/pip install --quiet sunpy==5.0.1
> $ sunpy/bin/python
> Python 3.11.6 (main, Oct  2 2023, 13:45:54) [Clang 15.0.0 (clang-1500.0.40.1)] on darwin
> Type "help", "copyright", "credits" or "license" for more information.
> >>> import sunpy.map
> Traceback (most recent call last):
>   File "<stdin>", line 1, in <module>
>   File "/private/tmp/sunpy/lib/python3.11/site-packages/sunpy/map/__init__.py", line 7, in <module>
>     from sunpy.map.mapbase import *
>   File "/private/tmp/sunpy/lib/python3.11/site-packages/sunpy/map/mapbase.py", line 15, in <module>
>     import matplotlib.pyplot as plt
> ModuleNotFoundError: No module named 'matplotlib'
> ```
>
> Maybe the `README.md` could mention that the `[map]` extra is required for the vignette? This also applies to the "Any additional setup required to use the package" criterion.

Yes this is a good point and I think at this point, only the coordinates module is considered "core" as you found out with map, you need the `map` extra.

This actually brings up two very interesting points about installation and how to provide that information.

Our readme, only suggests to install with conda and we do not really talk about installing sunpy via pip or "normal" Python at all until you get to https://docs.sunpy.org/en/stable/topic_guide/installation.html

We recommend that conda via miniforge is used to install sunpy which is why the readme is like that and that is how we frame it in our install documentation (https://docs.sunpy.org/en/stable/tutorial/installation.html).

I wonder what the best approach is, we rather avoid duplicating information into our readme, but the readme does end up at the landing page for the repo and the PyPI page.

I don't have any good ideas.

> 2. The API is _enormous_, so I can only spot-check if there's "Function Documentation: for all user-facing functions". But looks good.

Yeah, I tried to go through it myself and got lost midway.

> 3. The same applies to "Examples for all user-facing functions", however I find that `sunpy.util.expand_list` does not have an example. (This also applies to the "All functions have documentation and associated examples for use" criterion.)

I will see about adding some examples here and to other util parts. This part of the library was meant to be "developer" facing more than user facing (if one sees a difference there). The idea was that packages that use sunpy or maintainers of sunpy, could call upon these functions for internal code but not as "public" code. But this is not really how it works in practice.

> 4. Python packaging has been (sadly) a moving target, but this project seems to find the balance between `setup.cfg` and `pyproject.toml`. (The Python Packaging Authority has finally settled on moving away from `setup.py` and `setup.cfg` so perhaps later versions of `sunpy` can move all project metadata to `pyproject.toml` in a future version.)

Yeah, our setup.py I think is nearly empty and we plan to move to pyproject.toml but that will be held back as we are currently working on a package template and using cruft to auto re-template our packages. We want to avoid doing that manually.

> 5. "Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file". There is indeed one vignette in the `README.md` however I would _not_ call this a small package! 😆

While we do link to our example gallery in the readme, it's not very obvious.
I think there is much better formatting and layout work to be done to the readme to make it much better.

> 6. "Any performance claims of the software been confirmed" … the documentation doesn't make any mention of performance, so this criterion easily passes 😇.

Luckily we don't 😆

> 7. Running `flake8 --count sunpy` for "Code format is standard throughout package and follows PEP 8 guidelines" shows 501 violations

We have talked about adopting black to fix most of our code format but its contentious that we have yet to make a community decision over it. Hopefully that would fix most of these violations.

> 8. JOSS criteria: version 1.0.8 was reviewed by JOSS, but this version (5.0.1) does not have `paper.md`. Is version 5.0.1 going to be submitted to JOSS? If so, we'll need this file. If not, then no worries 😉

We did not plan to submit to JOSS as while it has been a long time since we applied. I don't think much has changed that JOSS would accept an updated paper.

> That's all I got! Thanks for the opportunity to review this.

Thank you for going through the package, sunpy has really grown and it isn't a straightforward package to review.
