+++ setup.cfg
@@ -234,6 +234,8 @@ filterwarnings =
     ignore:.*module is deprecated, as it was designed for internal use
     # This is raised when the VSO redirects and we do not want this to stop the CI
     ignore::sunpy.util.exceptions.SunpyConnectionWarning
+    # Can be removed when https://github.com/dateutil/dateutil/issues/1314 is resolved
+    ignore:datetime.datetime.utcfromtimestamp():DeprecationWarning
 
 [pycodestyle]
 max_line_length = 110
