### 2023-11-29 5c6b48a
Add pypi identifier for 3.12
