+++ setup.cfg
@@ -30,6 +30,7 @@ classifiers =
   Programming Language :: Python :: 3.9
   Programming Language :: Python :: 3.10
   Programming Language :: Python :: 3.11
+  Programming Language :: Python :: 3.12
   Topic :: Scientific/Engineering :: Physics
 
 [options]
