### 2023-12-01 53773d4
Update readme for pyopensci review
