+++ README.rst
@@ -2,74 +2,73 @@
 ``sunpy``
 *********
 
-|Latest Version| |DOI| |repostatus| |python|
-|ci| |codecov| |Docs|
+|Latest Version|
+|DOI|
+|repostatus|
+|python|
+|ci|
+|codecov|
+|Docs|
 |matrix|
 |Powered by NumFOCUS|
 
-.. |Latest Version| image:: https://img.shields.io/pypi/v/sunpy.svg
-   :target: https://pypi.python.org/pypi/sunpy/
+.. |Latest Version| image::https://anaconda.org/conda-forge/sunpy/badges/version.svg
+    :target: https://anaconda.org/conda-forge/sunpy
+    :alt: conda-forge - Latest Version
 .. |DOI| image:: https://zenodo.org/badge/2165383.svg
-   :target: https://zenodo.org/badge/latestdoi/2165383
+    :target: https://zenodo.org/badge/latestdoi/2165383
+    :alt: Zenodo - Latest DOI
 .. |matrix| image:: https://img.shields.io/matrix/sunpy:openastronomy.org.svg?colorB=%23FE7900&label=Chat&logo=matrix&server_fqdn=matrix.org
-   :target: https://app.element.io/#/room/#sunpy:openastronomy.org
+    :target: https://app.element.io/#/room/#sunpy:openastronomy.org
+    :alt: Matrix Chat Room
 .. |codecov| image:: https://codecov.io/gh/sunpy/sunpy/branch/main/graph/badge.svg
-   :target: https://codecov.io/gh/sunpy/sunpy
+    :target: https://codecov.io/gh/sunpy/sunpy
+    :alt: Code Coverage
 .. |Powered by NumFOCUS| image:: https://img.shields.io/badge/powered%20by-NumFOCUS-orange.svg?style=flat&colorA=E1523D&colorB=007D8A
-   :target: https://numfocus.org
+    :target: https://numfocus.org
+    :alt: Powered by NumFOCUS
 .. |CI| image:: https://github.com/sunpy/sunpy/actions/workflows/ci.yml/badge.svg?branch=main
-   :target: https://github.com/sunpy/sunpy/actions/workflows/ci.yml
+    :target: https://github.com/sunpy/sunpy/actions/workflows/ci.yml
+    :alt: Continuous Integration Status
 .. |Docs| image:: https://readthedocs.org/projects/sunpy/badge/?version=stable
-   :target: https://docs.sunpy.org/en/stable/?badge=stable
-   :alt: Documentation Status
+    :target: https://docs.sunpy.org/en/stable/?badge=stable
+    :alt: Stable Documentation
 .. |repostatus| image:: https://www.repostatus.org/badges/latest/active.svg
-   :alt: Project Status: Active – The project has reached a stable, usable state and is being actively developed.
-   :target: https://www.repostatus.org/#active
+    :target: https://www.repostatus.org/#active
+    :alt: Project Status: Active - The project has reached a stable, usable state and is being actively developed.
 .. |python| image:: https://img.shields.io/pypi/pyversions/sunpy
-   :alt: PyPI - Python Version
+    :alt: PyPI - Python Version
 
 ``sunpy`` is a Python software package that provides fundamental tools for accessing, loading and interacting with solar physics data in Python.
-It is the core library of the `SunPy Project <https://sunpy.org/>`__.
+It includes an interface for searching and downloading data from multiple data providers, data containers for image and time series data, commonly used solar coordinate frames and associated transformations, as well as other functionality needed for solar data analysis.
 
-For some examples of using ``sunpy`` see our `gallery <https://docs.sunpy.org/en/stable/generated/gallery/index.html>`__.
-To see the latest changes in ``sunpy`` see our `changelog <https://docs.sunpy.org/en/stable/whatsnew/changelog.html>`__.
[…]
