### 2023-12-01 #147 @cmarmo
Thank you @nutjob4life and @Septaris for your reviews!
Sunpy is a mature package and reviewing it is a lot of work.

> Authors are listed [here](https://github.com/sunpy/sunpy/blob/main/CITATION.cff); there is no individual email but a generic email (Google Group mail sunpy@googlegroups.com, plus the chat link). For me, it's fine.

@Septaris, this is fine for me too: you pointed out that the sunpy maintainers and the community around the package are quite responsive, so this not seems to be an issue.

I have changed the status of the issue as "awaiting changes": I understand some comments can be addressed while others are more long term suggestions.
Also, in order to allow the maintainers to reach a consensus about the modifications, @nabobalis, do you think a three weeks interval would be enough?
Thank you for your collaboration!
