### 2023-12-04 4eb77f1
Simplified POST request and added comments
