### 2023-12-04 2b0b624
Add a context manager for imposing a solar radius value
