### 2023-12-04 69c4eda
Add Clarification on setting Environment Variable for JSOC Email
