+++ changelog/7329.doc.rst
@@ -1,0 +1,1 @@
+Added Clarification on setting JSOC Email.
