### 2023-12-06 #147 @nabobalis
> I have changed the status of the issue as "awaiting changes": I understand some comments can be addressed while others are more long term suggestions. Also, in order to allow the maintainers to reach a consensus about the modifications, @nabobalis, do you think a three weeks interval would be enough? Thank you for your collaboration!

Yes that will be enough. Thank you very much again.
