### 2023-12-06 ada0d2b
Improved the wording around instllation and the several pages
