+++ docs/tutorial/installation.rst
@@ -5,7 +5,7 @@ Installation
 ************
 
 This is the first chapter in the sunpy tutorial, and by the end of it you should have a working installation of Python and ``sunpy``.
-For more info on different ways to install ``sunpy`` beyond the recommended way below, see :ref:`sunpy-topic-guide-installing`.
+For information on how to install ``sunpy`` beyond the recommended way below, see :ref:`sunpy-topic-guide-installing`.
 
 Installing Python
 =================
@@ -13,7 +13,8 @@ Installing Python
 There are many ways to install Python, but even if you have Python installed somewhere on your computer we recommend following these instructions anyway.
 That's because we will create a new Python environment.
 As well as containing a Python installation, this environment provides an isolated place to install Python packages (like ``sunpy``) without affecting any other current Python installation.
-If you already have Python and either ``conda`` or ``pip`` working you can skip the next section.
+If you already have Python and ``conda`` working you can skip the next section.
+However, if your working conda is an installation of Anaconda, we recommend you install miniforge as described below.
 
 Installing miniforge
 --------------------
@@ -57,7 +58,7 @@ Then select your platform to install miniforge:
 
         .. code-block:: console
 
-           bash Miniforge3-$(uname)-$(uname -m).sh
+            bash Miniforge3-$(uname)-$(uname -m).sh
 
         Once the installer has completed, close and reopen your terminal.
 
@@ -83,6 +84,7 @@ Now we will create and activate a new virtual environment to install ``sunpy`` i
     $ conda create --name sunpy
     # Only run the following two lines
     # if you have NOT installed miniforge or added conda-forge to your channels
+    # Please do not add these lines if you are using Anaconda
     $ conda config --add channels conda-forge
     $ conda config --set channel_priority strict
     $ conda activate sunpy
