### 2023-12-06 961982b
Point back to pypi as badge on conda-forge is wrong
