+++ README.rst
@@ -12,9 +12,9 @@
 |matrix|
 |Powered by NumFOCUS|
 
-.. |Latest Version| image::https://anaconda.org/conda-forge/sunpy/badges/version.svg
-    :target: https://anaconda.org/conda-forge/sunpy
-    :alt: conda-forge - Latest Version
+.. |Latest Version| image:: https://img.shields.io/pypi/v/sunpy.svg
+    :target: https://pypi.python.org/pypi/sunpy/
+    :alt: Latest Version
 .. |DOI| image:: https://zenodo.org/badge/2165383.svg
     :target: https://zenodo.org/badge/latestdoi/2165383
     :alt: Zenodo - Latest DOI
