+++ setup.cfg
@@ -96,6 +96,7 @@ all =
   %(timeseries)s
   %(visualization)s
 tests =
+  %(all)s
   hvpy>=1.0.1
   hypothesis>=6.0.0  # Included in pytest-astropy. 6.0 is the first version to support disabling function-scoped fixture warning
   jplephem  # For some coordinates tests
@@ -108,6 +109,7 @@ tests =
   pytest>=6.0
   %(spice)s
 docs =
+  %(all)s
   hvpy>=1.0.1
   packaging
   ruamel.yaml
