### 2023-12-06 2c3281f
Backport PR #7330 on branch 5.1 (Fix devdeps figure tests and getting the wrong filename) (#7332)
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
