+++ sunpy/tests/helpers.py
@@ -75,11 +75,11 @@ def get_hash_library_name():
     Generate the hash library name for this env.
     """
     import mpl_animators
-    version = mpl_animators.__version__
-    animators_version = "dev" if "+" in version else version.replace('.', '')
+
+    animators_version = "dev" if (("dev" in mpl_animators.__version__) or ("rc" in mpl_animators.__version__)) else mpl_animators.__version__.replace('.', '')
     ft2_version = f"{mpl.ft2font.__freetype_version__.replace('.', '')}"
-    mpl_version = "dev" if ("+" in mpl.__version__) or ("rc" in mpl.__version__) else mpl.__version__.replace('.', '')
-    astropy_version = "dev" if "dev" in astropy.__version__ else astropy.__version__.replace('.', '')
+    mpl_version = "dev" if (("dev" in mpl.__version__) or ("rc" in mpl.__version__)) else mpl.__version__.replace('.', '')
+    astropy_version = "dev" if (("dev" in astropy.__version__) or ("rc" in astropy.__version__)) else astropy.__version__.replace('.', '')
     return f"figure_hashes_mpl_{mpl_version}_ft_{ft2_version}_astropy_{astropy_version}_animators_{animators_version}.json"
 
 
