### 2023-12-06 #147 @nabobalis
Hello everyone, I have opened this PR: https://github.com/sunpy/sunpy/pull/7325

It changes the readme by deleting the install and example sections and just links out to our RTD pages that have that information. I feel like this might be against the spirit of the review and wanted to ask if doing this would be ok?

It also changes a few other minor things like using `[test]` will also pull in `all` among other changes that the reviewers spotted.
