+++ .github/workflows/scheduled_builds.yml
@@ -9,7 +9,7 @@ on:
     #        │ │ ┌───────── day of the month (1 - 31)
     #        │ │ │ ┌───────── month (1 - 12 or JAN-DEC)
     #        │ │ │ │ ┌───────── day of the week (0 - 6 or SUN-SAT)
-    - cron: '0 7 * * *'  # Every day at 07:00 UTC
+    - cron: "0 7 * * *" # Every day at 07:00 UTC
 
 jobs:
   dispatch_workflows:
