+++ CITATION.cff
@@ -1,79 +1,79 @@
-cff-version: '1.1.0'
-message: 'Please cite the following work when using this software.'
+cff-version: "1.1.0"
+message: "Please cite the following work when using this software."
 authors:
-  - name: 'The SunPy Community'
-  - family-names: 'Barnes'
-    given-names: 'Will T.'
-  - family-names: 'Bobra'
-    given-names: 'Monica G.'
-  - family-names: 'Christe'
-    given-names: 'Steven D.'
-  - family-names: 'Freij'
-    given-names: 'Nabil'
-  - family-names: 'Hayes'
-    given-names: 'Laura A.'
-  - family-names: 'Ireland'
-    given-names: 'Jack'
-  - family-names: 'Mumford'
-    given-names: 'Stuart'
-  - family-names: 'Perez-Suarez'
-    given-names: 'David'
-  - family-names: 'Ryan'
-    given-names: 'Daniel F.'
-  - family-names: 'Shih'
-    given-names: 'Albert Y.'
-  - family-names: 'Chanda'
-    given-names: 'Prateek'
-  - family-names: 'Glogowski'
-    given-names: 'Kolja'
-  - family-names: 'Hewett'
-    given-names: 'Russell'
-  - family-names: 'Hughitt'
-    given-names: 'V. Keith'
-  - family-names: 'Hill'
-    given-names: 'Andrew'
-  - family-names: 'Hiware'
-    given-names: 'Kaustubh'
-  - family-names: 'Inglis'
-    given-names: 'Andrew'
-  - family-names: 'Kirk'
-    given-names: 'Michael S. F.'
-  - family-names: 'Konge'
-    given-names: 'Sudarshan'
-  - family-names: 'Mason'
-    given-names: 'James Paul'
-  - family-names: 'Maloney'
-    given-names: 'Shane Anthony'
-  - family-names: 'Murray'
-    given-names: 'Sophie A.'
-  - family-names: 'Panda'
-    given-names: 'Asish'
-  - family-names: 'Park'
-    given-names: 'Jongyeob'
-  - family-names: 'Pereira'
-    given-names: 'Tiago M. D.'
-  - family-names: 'Reardon'
-    given-names: 'Kevin'
-  - family-names: 'Savage'
-    given-names: 'Sabrina'
-  - family-names: 'Sipőcz'
-    given-names: 'Brigitta M.'
-  - family-names: 'Stansby'
-    given-names: 'David'
-  - family-names: 'Jain'
-    given-names: 'Yash'
-  - family-names: 'Taylor'
-    given-names: 'Garrison'
-  - family-names: 'Yadav'
-    given-names: 'Tannmay'
-  - family-names: 'Rajul'
-  - family-names: 'Dang'
-    given-names: 'Trung Kien'
-doi: '10.3847/1538-4357/ab4f7a'
+  - name: "The SunPy Community"
+  - family-names: "Barnes"
+    given-names: "Will T."
+  - family-names: "Bobra"
+    given-names: "Monica G."
+  - family-names: "Christe"
+    given-names: "Steven D."
+  - family-names: "Freij"
+    given-names: "Nabil"
+  - family-names: "Hayes"
+    given-names: "Laura A."
+  - family-names: "Ireland"
+    given-names: "Jack"
+  - family-names: "Mumford"
+    given-names: "Stuart"
+  - family-names: "Perez-Suarez"
+    given-names: "David"
+  - family-names: "Ryan"
+    given-names: "Daniel F."
+  - family-names: "Shih"
+    given-names: "Albert Y."
+  - family-names: "Chanda"
+    given-names: "Prateek"
+  - family-names: "Glogowski"
+    given-names: "Kolja"
[…]
