+++ .github/workflows/changelog_bot.yml
@@ -7,8 +7,8 @@ on:
       - main
 
 env:
-  BOT_USER_NAME: 'SunPyBot'
-  BOT_USER_EMAIL: 'bot@sunpy.org'
+  BOT_USER_NAME: "SunPyBot"
+  BOT_USER_EMAIL: "bot@sunpy.org"
 
 jobs:
   update-changelog:
@@ -23,7 +23,7 @@ jobs:
       - name: Setup Python 3.x
         uses: actions/setup-python@v4
         with:
-          python-version: '3.x'
+          python-version: "3.x"
       - name: Find and write changelog
         run: python tools/write_changelog_entry.py
         env:
@@ -33,10 +33,10 @@ jobs:
       - name: Commit changelog
         uses: stefanzweifel/git-auto-commit-action@v5
         with:
-          commit_message: '[skip ci] Add changelog'
-          add_options: '-A'
+          commit_message: "[skip ci] Add changelog"
+          add_options: "-A"
           disable_globbing: true
-          file_pattern: 'changelog/*.rst'
+          file_pattern: "changelog/*.rst"
           commit_user_name: ${{ env.BOT_USER_NAME }}
           commit_user_email: ${{ env.BOT_USER_EMAIL }}
           commit_author: ${{ env.BOT_USER_NAME }} <${{ env.BOT_USER_EMAIL }}>

[…]
