### 2023-12-07 e000af8
Remove sunpy.database (#7320)
* Remove sunpy.database

* Add changelog+what's new

* Fix changelog links
