+++ sunpy/database/tests/test_attrs.py
@@ -1,485 +1,0 @@
-# Author: Simon Liedtke <liedtke.simon@googlemail.com>
-#
-# This module was developed with funding provided by
-# the Google Summer of Code (2013).
-
-from datetime import datetime
-
-import pytest
-
-import astropy.units as u
-
-from sunpy.database import tables
-from sunpy.database.attrs import DownloadTime, FitsHeaderEntry, Path, Starred, Tag, walker
-from sunpy.database.database import Database
-from sunpy.net import attrs as a
-from sunpy.net import vso
-from sunpy.net.attr import AttrAnd, AttrOr, DummyAttr
-
-
-@pytest.fixture
-def obj():
-    return object()
-
-
-@pytest.fixture
-def session():
-    database = Database('sqlite:///:memory:')
-    for i in range(1, 11):
-        entry = tables.DatabaseEntry()
-        database.add(entry)
-        # every entry has a fake download time of 2005-06-15 i:00:00
-        database.edit(entry, download_time=datetime(2005, 6, 15, i))
-        # every second entry gets starred
-        if i % 2 == 0:
-            database.star(entry)
-        # every third entry is stored in the path /tmp
-        if i % 3 == 0:
-            database.edit(entry, path='/tmp')
-        # every fifth entry gets the tag 'foo'
-        if i % 5 == 0:
-            database.tag(entry, 'foo')
-    # the last entry gets the FITS header entry INSTRUME=EIT
-    entry.fits_header_entries.append(tables.FitsHeaderEntry('INSTRUME', 'EIT'))
-    database.commit()
-    return database.session
-
-
-@pytest.fixture
-def vso_session():
-    client = vso.VSOClient()
-    qr = client.search(
-        a.Time((2011, 9, 20, 1), (2011, 9, 20, 2)),
-        a.Instrument.rhessi, response_format="legacy")
-    entries = tables.entries_from_query_result(qr)
-    database = Database('sqlite:///:memory:')
-    for entry in entries:
-        database.add(entry)
-    database.commit()
-    return database.session
-
-
-def test_starred_nonzero():
-    assert Starred()
-
-
-def test_starred_invert():
-    assert not ~Starred()
-
-
-def test_starred_and_different_types(obj):
-    assert Starred() & obj == AttrAnd([Starred(), obj])
-
-
-def test_starred_and_same_types():
-    assert ~Starred() & ~Starred() == ~Starred()
-    assert ~Starred() & Starred() == ~Starred()
-    assert Starred() & ~Starred() == ~Starred()
-    assert Starred() & Starred() == Starred()
-
-
-def test_starred_or_different_types(obj):
-    assert Starred() | obj == AttrOr([Starred(), obj])
-
-
-def test_starred_or_same_types():
-    assert ~Starred() | ~Starred() == ~Starred()
-    assert ~Starred() | Starred() == Starred()
-    assert Starred() | ~Starred() == Starred()
-    assert Starred() | Starred() == Starred()
-
-
-def test_starred_equality():
-    assert Starred() == Starred()
-    assert Starred() != ~Starred()
-    assert ~Starred() != Starred()
-    assert ~Starred() == ~Starred()
-
-
-def test_starred_repr():
-    assert repr(Starred()) == '<Starred()>'
-    assert repr(~Starred()) == '<~Starred()>'
-
-
-def test_tag_repr():
-    assert repr(Tag('foo')) == "<Tag('foo')>"
-    assert repr(~Tag('foo')) == "<~Tag('foo')>"
-
-
-def test_path_repr():
-    assert repr(Path('/tmp')) == "<Path('/tmp')>"
-    assert repr(~Path('/tmp')) == "<~Path('/tmp')>"
-
-
-def test_downloadtime_repr():
[…]
