+++ sunpy/util/tests/test_sysinfo.py
@@ -24,7 +24,6 @@ def test_find_dependencies():
                                                      'packaging',
                                                      'requests',
                                                      'dask',
-                                                     'sqlalchemy',
                                                      'scikit-image',
                                                      'scipy',
                                                      'glymur',
@@ -49,7 +48,6 @@ def test_missing_dependencies_by_extra():
                                                    'asdf',
                                                    'required',
                                                    'dask',
-                                                   'database',
                                                    'dev',
                                                    'docs',
                                                    'docs-gallery',
@@ -62,7 +60,7 @@ def test_missing_dependencies_by_extra():
                                                    'timeseries',
                                                    'visualization'])
     missing = missing_dependencies_by_extra(exclude_extras=["all"])
-    assert sorted(list(missing.keys())) == sorted(['asdf', 'required', 'dask', 'database', 'dev', 'docs',
+    assert sorted(list(missing.keys())) == sorted(['asdf', 'required', 'dask', 'dev', 'docs',
                                                    'docs-gallery', 'image', 'jpeg2000', 'map', 'net',
                                                    'spice', 'tests', 'timeseries', 'visualization'])
 
