+++ changelog/7320.removal.rst
@@ -1,0 +1,1 @@
+``sunpy.database`` has been removed.
