+++ sunpy-dev-env.yml
@@ -24,7 +24,6 @@ dependencies:
   - scikit-image
   - scipy
   - setuptools
-  - sqlalchemy
   - tqdm
   - zeep
 

[…]
