### 2023-12-07 dfdf213
Removed HelioviewerClient
