+++ sunpy/net/helioviewer.py
@@ -1,459 +1,0 @@
-"""
-This module provides a wrapper around the Helioviewer API.
-"""
-import json
-import codecs
-import urllib
-from pathlib import Path
-from collections import OrderedDict
-from urllib.parse import urljoin
-from urllib.request import urlopen
-
-from astropy.utils.decorators import lazyproperty
-
-import sunpy
-import sunpy.util.parfive_helpers as parfive
-from sunpy import log
-from sunpy.time import parse_time
-from sunpy.util.decorators import deprecated
-from sunpy.util.util import partial_key_match
-from sunpy.util.xml import xml_to_dict
-
-__all__ = ['HelioviewerClient']
-
-HELIOVIEWER_API_URLS = [
-    "https://api.helioviewer.org/",
-    "https://helioviewer-api.ias.u-psud.fr/",
-]
-
-
-def check_connection(url):
-    try:
-        resp = urlopen(urljoin(url, "/v2/getDataSources/"))
-        assert resp.getcode() == 200
-        assert isinstance(json.loads(resp.read()), dict)
-        return url
-    except Exception as e:
-        log.debug(f"Unable to connect to {url}:\n {e}")
-        log.info(f"Connection to {url} failed. Retrying with different url.")
-    return None
-
-
-@deprecated(since="4.1", alternative="hvpy - https://hvpy.readthedocs.io/en/stable")
-class HelioviewerClient:
-    """Helioviewer.org Client"""
-
-    def __init__(self, url=None):
-        """
-        Parameters
-        ----------
-        url : `str`
-            Default URL points to the Helioviewer API.
-        """
-        if url is None:
-            for url in HELIOVIEWER_API_URLS:
-                if check_connection(url):
-                    break
-
-        if url is None:
-            raise ValueError("No online helioviewer API can be found.")
-
-        self._api = url
-
-    @lazyproperty
-    def data_sources(self):
-        """
-        We trawl through the return from ``getDataSources`` to create a clean
-        dictionary for all available sourceIDs.
-
-        Here is a list of all of them: https://api.helioviewer.org/docs/v2/#appendix_datasources
-        """
-        data_sources_dict = dict()
-        datasources = self.get_data_sources()
-        for name, observ in datasources.items():
-            # TRACE only has measurements and is thus nested once
-            if name == "TRACE":
-                for instr, params in observ.items():
-                    data_sources_dict[(name, None, None, instr)] = params['sourceId']
-            else:
-                for inst, detect in observ.items():
-                    for wavelength, params in detect.items():
-                        if 'sourceId' in params:
-                            data_sources_dict[(name, inst, None, wavelength)] = params['sourceId']
-                        else:
-                            for wave, adict in params.items():
-                                data_sources_dict[(name, inst, wavelength, wave)
-                                                  ] = adict['sourceId']
-
-        # Sort the output for printing purposes
-        return OrderedDict(sorted(data_sources_dict.items(), key=lambda x: x[1]))
-
-    def get_data_sources(self):
-        """
-        Return a hierarchical dictionary of the available datasources on helioviewer.org.
-
-        This uses ``getDataSources`` from the Helioviewer API.
-
-        Returns
-        -------
-        out : `dict`
-            A dictionary containing meta-information for each data source that Helioviewer supports.
-        """
-        params = {"action": "getDataSources"}
-        return self._get_json(params)
-
-    def get_closest_image(self, date, observatory=None, instrument=None,
-                          detector=None, measurement=None, source_id=None):
-        """
-        Finds the closest image available for the specified source and date.
-        **This does not download any file.**
-
-        This uses `getClosestImage <https://api.helioviewer.org/docs/v2/#OfficialClients>`_
-        from the Helioviewer API.
-
[…]
