+++ sunpy/map/tests/test_reproject_to.py
@@ -13,7 +13,6 @@ from astropy.wcs import WCS
 
 import sunpy.map
 from sunpy.tests.helpers import figure_test
-from sunpy.util.exceptions import SunpyDeprecationWarning
 
 
 @pytest.fixture
@@ -120,11 +119,3 @@ def test_return_footprint(aia171_test_map, hpc_header):
 def test_invalid_inputs(aia171_test_map, hpc_header):
     with pytest.raises(ValueError, match="The specified algorithm must be one of"):
         aia171_test_map.reproject_to(hpc_header, algorithm='something')
-
-
-def test_deprecated_positional_args(aia171_test_map, hpc_header):
-    with pytest.warns(SunpyDeprecationWarning, match=r"Pass algorithm=interpolation as keyword args"):
-        aia171_test_map.reproject_to(hpc_header, 'interpolation')
-
-    with pytest.warns(SunpyDeprecationWarning, match=r"Pass algorithm=interpolation, return_footprint=True as keyword args"):
-        aia171_test_map.reproject_to(hpc_header, 'interpolation', True)
