### 2023-12-07 6f603dc
Removal of response_format for vso
