+++ sunpy/net/vso/table_response.py
@@ -55,17 +55,11 @@ class VSOQueryResponseTable(QueryResponseTable):
     size_column = 'Size'
 
     @classmethod
-    def from_zeep_response(cls, response, *, client, _sort=True):
+    def from_zeep_response(cls, response, *, client):
         """
         Construct a table response from the zeep response.
         """
-        # _sort is a hack to be able to convert from a legacy QueryResponse to
-        # a table response.
-        if _sort:
-            records = iter_sort_response(response)
-        else:
-            records = response
-
+        records = iter_sort_response(response)
         data = []
         for record in records:
             row = defaultdict(lambda: None)
