+++ sunpy/net/fido_factory.py
@@ -479,9 +479,6 @@ class UnifiedDownloaderFactory(BasicRegistrationFactory):
         for client in candidate_widget_types:
             tmpclient = client()
             kwargs = dict()
-            # Handle the change in response format in the VSO
-            if isinstance(tmpclient, vso.VSOClient):
-                kwargs = dict(response_format="table")
             results.append(tmpclient.search(*query, **kwargs))
 
         # This method is called by `search` and the results are fed into a
