### 2023-12-11 #147 @Septaris
Here are my thoughts and suggestions based on your comments and the merge request:

- Conda Badge: The Conda badge is still missing.
- Continuous Integration: If the current behavior regarding continuous integration is intentional, then it's fine.
- Citation Link: The link has been renamed/updated. ✅
- Tests Documentation: Regarding the Matplotlib backend (MPLBACKEND = agg), it would be nice to include this information in the pytest-related documentation.
- Package supports modern versions of Python: Support for version 3.8 will be dropped in a few months, so it's not a problem. However, support for version 3.12 seems required.

About the **Installation Instructions**:

1. The review grid is strict and indicates the need to integrate installation instructions for the development version and vignette(s) directly into the README. However, from my perspective, it's acceptable if you propose a more streamlined approach by providing links to the gallery and standard/development installation instructions in the README. It seems reasonable and can prevent information duplication. I'll let our editor, @cmarmo, decide.

2. Now, on the installation documentation content, specifically `pip` vs `conda`: From my experience, installing `sunpy` is easier via pip (`sunpy[all]`) than via conda, especially on a standard operating system where the `wheels` are available. I'm afraid that beginners who use `pip` (which remains the Python standard) might get confused at the first mention of `conda`. Therefore, I recommend the following structure:

   - A standard installation procedure (with Conda and Pip sections).
   - An installation procedure in development mode (again with Conda and Pip sections).
   - A readme that points to the two installation procedures.

    Note: this is only a suggestion and is not a blocking point for the review. I will let you decide based on the needs/feedback of the community.

And now, you still need to ensure compliance with the PEP 8 rules... Good luck! 😄
