### 2023-12-11 2f42c40
Added note about warning when running a figure test without the MPLBACKEND
