+++ docs/dev_guide/contents/tests.rst
@@ -77,6 +77,13 @@ If a test errors, you can use ``pdb`` to create a debugging session at the momen
 
     $ pytest --pdb
 
+If you see mention of::
+
+    UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
+
+You will have to either export ``MPLBACKEND = agg`` as an environmental variable or pass it as a command line pre-argument to ``pytest``.
+This comes from the figure tests (see below).
+
 ``self_test``
 =============
 
@@ -106,7 +113,8 @@ If this is the case, you will see the following when you run the test suite::
     You do not have all the required dependencies installed to run the sunpy test suite.
     If you want to run the sunpy tests install the 'tests' extra with `pip install "sunpy[tests]"`
 
-This does not mean sunpy is broken, but you need a "complete" installation of sunpy to run the entire test suite.
+This does not mean sunpy is broken, but you will need to install the extra packages to ensure a "complete" installation of sunpy and run the entire test suite.
+It is quite likely that you will run into not having the tests dependencies installed.
 
 Remote data
 -----------
