### 2023-12-11 e818975
Do not run the figure tests by default
