### 2023-12-12 6e134fb
Update mapbase.py
try...except used to provide a useful error message to users.
