### 2023-12-12 5231691
Update sunpy/map/mapbase.py
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
