### 2023-12-12 6d9572d
Update sunpy/map/mapbase.py
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
