### 2023-12-12 ca09ef8
Update sunpy/map/mapbase.py
