### 2023-12-12 4808136
Update sunpy/map/mapbase.py
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
