### 2023-12-12 32e5d93
Update sunpy/map/mapbase.py
