### 2023-12-13 03a5059
added unit test for mapbase file change test_mapbase.py
changed unit test according to the try..except added to the padding
