+++ sunpy/map/tests/test_mapbase.py
@@ -1036,22 +1036,12 @@ def test_rotate_with_incompatible_missing_dtype_warning():
                      observer='earth', frame=sunpy.coordinates.Helioprojective)
     header = sunpy.map.make_fitswcs_header(data, coord)
     test_map = sunpy.map.Map(data, header)
+    
     with pytest.warns(SunpyUserWarning,
                       match="The specified `missing` value is not an integer, but the data "
-                      "array is of integer type, so the output may be strange."):
+                            "array is of integer type, so the output may be strange."):
         test_map.rotate(order=3, missing=np.nan)
 
-
-@pytest.mark.skipif(version.parse(np.__version__) <= version.parse("1.20.0"),
-                    reason="Numpy >= 1.20.0 doesn't allow NaN to int conversion")
-def test_rotate_with_incompatible_missing_dtype_error():
-    data = np.arange(0, 100).reshape(10, 10)
-    coord = SkyCoord(0 * u.arcsec, 0 * u.arcsec, obstime='2013-10-28',
-                     observer='earth', frame=sunpy.coordinates.Helioprojective)
-    header = sunpy.map.make_fitswcs_header(data, coord)
-    test_map = sunpy.map.Map(data, header)
-    with pytest.raises(ValueError, match="cannot convert float NaN to integer"):
-        test_map.rotate(order=3, missing=np.nan)
 
 
 def test_rotate_crpix_zero_degrees(generic_map):
