### 2023-12-13 36c522e
unit test added to make sure we hit the error in test_mapbase.py
