### 2023-12-16 1a49f90
Update test_mapbase.py
added unit test with raising a warning rather than an error in padding
