### 2023-12-17 9286699
Update sunpy/map/mapbase.py
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
