### 2023-12-18 9ea5242
added a common unit test for both error and warning in numpy padding.
