### 2023-12-18 85ec23b
new_data variable introduced for except method aswell
