### 2023-12-19 37e036f
if condition for numpy <1.20.0 removed
