### 2023-12-19 cd761a9
raising value error after except in mapbase.py
