### 2023-12-19 1f16807
Update test_mapbase.py
