+++ sunpy/map/tests/test_mapbase.py
@@ -1025,7 +1025,8 @@ def test_rotate(aia171_test_map):
     aia171_test_map_crop_rot = aia171_test_map_crop.rotate(60 * u.deg)
     assert aia171_test_map_crop.data.shape[0] > aia171_test_map_crop.data.shape[1]
     assert aia171_test_map_crop_rot.data.shape[0] < aia171_test_map_crop_rot.data.shape[1]
-    
+
+
 def test_rotate_with_incompatible_missing_dtype_error():
     data = np.arange(0, 100).reshape(10, 10)
     coord = SkyCoord(0 * u.arcsec, 0 * u.arcsec, obstime='2013-10-28',
@@ -1035,6 +1036,7 @@ def test_rotate_with_incompatible_missing_dtype_error():
     with pytest.raises(ValueError, match="Unable to pad the array most likely due to the default `missing` keyword as it is set to NaN. "
                                          "Change the `missing` keyword to a different value."):
         test_map.rotate(order=3, missing=np.nan)
+
 
 def test_rotate_crpix_zero_degrees(generic_map):
     # Rotating a map by zero degrees should not change the location of the reference pixel at all
