+++ sunpy/map/mapbase.py
@@ -1710,7 +1710,7 @@ class GenericMap(NDData):
             error_message = "Unable to pad the array most likely due to the default `missing` keyword as it is set to NaN. "
             "Change the `missing` keyword to a different value."
             raise ValueError(error_message)
-            
+
         # All of the following pixel calculations use a pixel origin of 0
         pixel_array_center = (np.flipud(new_data.shape) - 1) / 2.0
         pixel_rotation_center = u.Quantity(self.reference_pixel).value + [pad_x, pad_y]
