### 2023-12-20 a1065f8
Update sunpy/map/tests/test_mapbase.py
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
