### 2023-12-20 031802c
Update sunpy/map/mapbase.py
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
