### 2023-12-20 db64094
Update sunpy/map/mapbase.py
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
