+++ sunpy/map/mapbase.py
@@ -1699,8 +1699,7 @@ class GenericMap(NDData):
         unpad_x = -np.min((diff[0], 0))
         unpad_y = -np.min((diff[1], 0))
 
-        # Numpy 1.20+ prevents np.pad from padding with NaNs in integer arrays
-        # Before it would be cast to 0, but now it raises an error.
+        # Raise an informative error message if trying to pad an integer array with NaNs
     if (pad_x > 0 or pad_y > 0) and issubclass(self.data.dtype.type, numbers.Integral) and (missing % 1 != 0):
             raise ValueError("The underlying data is integers, but the fill value for missing "
                              "pixels cannot be cast to an integer, which is the case for the "
