### 2023-12-21 b734df6
logic fixed and test case updated for mapbase.py
