+++ sunpy/map/tests/test_mapbase.py
@@ -1035,7 +1035,7 @@ def test_rotate_with_incompatible_missing_dtype_error():
     test_map = sunpy.map.Map(data, header)
     with pytest.raises(ValueError, match="The underlying data is integers, but the fill value for "
                                          "missing pixels cannot be cast to an integer"):
-        test_map.rotate(order=3, missing=np.nan)
+        test_map.rotate(angle=45 * u.deg, missing=np.nan, order=3)
 
 
 def test_rotate_crpix_zero_degrees(generic_map):
