### 2023-12-23 #147 @cmarmo
Hello everybody,
touching base as the three weeks has passed and I didn't realize the deadline was bringing us just before Christmas holidays! 🎄
To sum up: the pull requests addressing the reviewer comments are https://github.com/sunpy/sunpy/pull/7325 and  https://github.com/sunpy/sunpy/pull/7343, which still need approval from the SunPy community.
I'd like to make the vacation break official: I'm going to check in around January 10th so we can finalize the discussion.
Let me know if this is ok for all the people involved.

Thanks and happy Christmas and end of the year!
