### 2023-12-25 e2a0942
Raise our own error when trying to rotate a map of integer data while using a `missing` value of NaN (#7344)
* Update mapbase.py

try...except used to provide a useful error message to users.

* Update sunpy/map/mapbase.py

Co-authored-by: Nabil Freij <nabil.freij@gmail.com>

* Update sunpy/map/mapbase.py

Co-authored-by: Nabil Freij <nabil.freij@gmail.com>

* Update sunpy/map/mapbase.py

* Update sunpy/map/mapbase.py

Co-authored-by: Nabil Freij <nabil.freij@gmail.com>

* Update sunpy/map/mapbase.py

* added unit test for mapbase file change test_mapbase.py

changed unit test according to the
[…]
