### 2023-12-25 328c990
Backport PR #7344: Raise our own error when trying to rotate a map of integer data while using a `missing` value of NaN
