### 2023-12-25 e95869c
Backport PR #7344 on branch 5.0 (Raise our own error when trying to rotate a map of integer data while using a `missing` value of NaN) (#7346)
Co-authored-by: Manit Singh <79140607+NucleonGodX@users.noreply.github.com>
