### 2023-12-26 5201d79
Added an example of plotting a rectangle on a map with a rotation angle relative to the axes
