### 2023-12-26 e96c7d2
reformatted the added file based on other examples, added changelog, fixed pre-commits
