### 2023-12-27 a81e557
fixed second plot not visible in gallery
