+++ changelog/7342.trivial.rst
@@ -1,0 +1,1 @@
+`~sunpy.net.jsoc.attrs.Notify` checks for a valid input value.
