### 2023-12-28 5bfd931
Fix test and add whatsnew
