+++ sunpy/timeseries/sources/tests/test_eve.py
@@ -1,11 +1,10 @@
-import re
 
 import pytest
 
 import sunpy.timeseries
 from sunpy.data.test import get_test_filepath
 from sunpy.tests.helpers import figure_test
-from sunpy.util.exceptions import SunpyDeprecationWarning, SunpyUserWarning
+from sunpy.util.exceptions import SunpyUserWarning
 
 esp_filepath = get_test_filepath('eve_l1_esp_2011046_00_truncated.fits')
 eve_filepath = get_test_filepath('EVE_L0CS_DIODES_1m_truncated.txt')
@@ -36,11 +35,6 @@ def test_esp_plot_column(esp_test_ts):
     assert '0.1-7nm' in axes[0].get_ylabel()
     assert '18nm' in axes[1].get_ylabel()
     assert '36nm' in axes[2].get_ylabel()
-
-
-def test_esp_peek_keyword_deprecation(esp_test_ts):
-    with pytest.warns(SunpyDeprecationWarning, match=re.escape('Pass title= as keyword args')):
-        esp_test_ts.peek('')
 
 
 @figure_test
