### 2023-12-28 f69323c
Applied suggestions from code review
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
