+++ examples/plotting/plot_rotated_rectangle.py
@@ -1,7 +1,7 @@
 """
-============================
+====================================
 Drawing a rotated rectangle on a map
-============================
+====================================
 
 This example will demonstrate how to draw a rectangle that can be rotated relative
 to the axes on a map using :meth:`~sunpy.map.GenericMap.draw_quadrangle`.
@@ -15,19 +15,19 @@ import sunpy.data.sample
 import sunpy.map
 
 ################################################################################
-# Let's start with a sample AIA image.
-
-aia_map = sunpy.map.Map(sunpy.data.sample.AIA_171_IMAGE)
+# Let's start with a sample image of AIA 171.
 
 # sphinx_gallery_defer_figures
+
+aia_map = sunpy.map.Map(sunpy.data.sample.AIA_171_IMAGE)
 
 fig = plt.figure(figsize=(5, 5))
 ax = fig.add_subplot(projection=aia_map)
 aia_map.plot(axes=ax, clip_interval=(1, 99.99) * u.percent)
 
 ################################################################################
-# Define the rotation angle and center coordinate of the rectangle
-# Width and height of the rectangle
+# Define the rotation angle and center coordinate of the rectangle,
+# as well as the width and height in physical units.
 
 # sphinx_gallery_defer_figures
 
@@ -37,9 +37,12 @@ width = 400 * u.arcsec
 height = 300 * u.arcsec
 
 ################################################################################
-# Create a SkyOffsetFrame with the specified rotation and center
+# To specify the rotation and center, we will use `SkyOffsetFrame`.
 # Define the corner coordinates in the SkyCoord frame
 # Transform corner coordinates to the original map's coordinate frame
+
+# sphinx_gallery_defer_figures
+
 offset_frame = SkyOffsetFrame(origin=center_coord, rotation=rotation_angle)
 corner_bottom_left = SkyCoord(lon=-width / 2, lat=-height / 2, frame=offset_frame)
 corner_top_right = SkyCoord(lon=width / 2, lat=height / 2, frame=offset_frame)
@@ -47,7 +50,8 @@ corner_bottom_left = corner_bottom_left.transform_to(aia_map.coordinate_frame)
 corner_top_right = corner_top_right.transform_to(aia_map.coordinate_frame)
 
 ################################################################################
-# Draw the rotated rectangle
+# Finally, we will draw the rotated rectangle.
+
 aia_map.draw_quadrangle(
     bottom_left=corner_bottom_left,
     top_right=corner_top_right,
