### 2023-12-28 2a53979
Tests plus figure hashes
