+++ changelog/7307.trivial.rst
@@ -1,3 +1,11 @@
-Progress from :meth:`sunpy.net.jsoc.JSOCClient.fetch` is now reported through the logger.
-Pinned the ``drms`` requirement to ``>= 0.7`` to account for this breaking API change in ``drms``.
-Note that this change supersedes that in `PR #7308 <https://github.com/sunpy/sunpy/pull/7308>`_.
+:meth:`sunpy.net.jsoc.JSOCClient.fetch` called `drms` API that passed a ``progress`` keyword which added extra print statements to the console.
+This has been removed in `drms` 0.7.0, which had breaking API changes within this release.
+As a result, we increased the minimum required version of `drms` to 0.7.0.
+
+This specifically refers to the following information that was printed to the console:
+
+``"Export request pending. [id=X, status=X]"``
+``"Waiting for X seconds..."``
+``"Request not found on server, X retries left."``
+
+These were handled by `drms` and are now logging messages.
