### 2023-12-28 537e1ca
changed the image size to default in plot_rotated_rectangle.py and added the sphinx interlink to example file
