+++ changelog/7348.feature.rst
@@ -1,1 +1,1 @@
-Add an example of plotting a rectangle on a map with a rotation angle relative to the axes.
+Add an example of plotting a rectangle on a map with a rotation angle relative to the axes :ref:`sphx_glr_generated_gallery_plotting_plot_rotated_rectangle.py`
