### 2023-12-28 2332d3b
Apply suggestions from code review
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
