### 2023-12-29 3fc05a6
Update setup.cfg
