+++ setup.cfg
@@ -70,7 +70,7 @@ map =
   scipy>=1.7.0,!=1.10.0
 net =
   beautifulsoup4>=4.8.0
-  drms>=0.7.0
+  drms>=0.7.1
   python-dateutil>=2.8.0
   tqdm>=4.32.1
   zeep>=3.4.0
