### 2023-12-29 0d5c583
Update changelog/7348.feature.rst
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
