### 2023-12-29 8a64c13
Add function to call specfic submodule unit tests
