+++ tox.ini
@@ -116,6 +116,8 @@ commands =
 [testenv:base_deps]
 description = Check the test suite does not fail if all optional dependencies are missing
 extras =
+allowlist_externals=
+    pytest
 deps =
     astropy
     numpy
@@ -123,7 +125,7 @@ deps =
 commands =
     python -c "import sunpy; sunpy.self_test()"
     # Test that we can import subpackages which don't have specific extras
-    # TODO: Actually run the tests for these modules
+    # and that the tests don't fail if optional dependencies are missing
     python -c "import sunpy.coordinates"
     python -c "import sunpy.data"
     python -c "import sunpy.io"
@@ -131,3 +133,5 @@ commands =
     python -c "import sunpy.sun"
     python -c "import sunpy.time"
     python -c "import sunpy.util"
+    python -m pip install pytest # Needed for the below command
+    python -c "from sunpy.tests.self_test import _test_base_deps; _test_base_deps()"
