+++ setup.cfg
@@ -69,6 +69,7 @@ map =
 net =
   beautifulsoup4>=4.8.0
   drms>=0.6.1,<0.7.0
+  parfive[ftp]>=2.0.0
   python-dateutil>=2.8.0
   tqdm>=4.32.1
   zeep>=3.4.0
