+++ docs/how_to/fix_map_metadata.rst
@@ -1,13 +1,15 @@
 .. _sunpy-how-to-fix-map-metadata:
 
-********************
-Fix Metadata for map
-********************
+*************************
+Fixing incorrect metadata
+*************************
 
-There will be times where you will come across either a FITS files with incorrect metadata, or metadata that is not understood by ``sunpy``.
-To load these files in to `sunpy.map.Map` you will need to fix the metadata before hand.
+There will be times where you will come across a FITS files with either incorrect, missing or unparsable metadata and reading these files into `~sunpy.map.Map` will cause an error.
+Therefore, to load these files into a `~sunpy.map.Map`, you will need to correct the metadata before hand.
 
-This will heavily depend on what metadata is incorrect, for example if the units in the FITS header are incorrect, (which will raise a metadata warning) you can fix this by changing the ``cunit1`` and ``cunit2`` keywords to the correct units:
+This will heavily depend on what metadata requires changing, unforutanlly it is near impossible to have a guide for every possible keyword.
+Below we have an example where the units in the FITS header are incorrect, this is controlled by the ``cunit1`` and ``cunit2`` keywords in a FITS header.
+So we will change those values and then load the file into a `~sunpy.map.Map`.
 
 .. code-block:: python
 
@@ -23,5 +25,4 @@ This will heavily depend on what metadata is incorrect, for example if the units
     >>> header['cunit2'] = 'arcsec'  # doctest: +REMOTE_DATA
     >>> updated_map = Map(data, header)  # doctest: +REMOTE_DATA
 
-This would apply for the observer_coordinate, wavelength, exposure time, etc.
-If you are unsure what the correct metadata should be, you can check the `FITS standard <https://fits.gsfc.nasa.gov/fits_standard.html>`__.
+This applies for any FITS standard keyword, which you can find in the `FITS standard <https://fits.gsfc.nasa.gov/fits_standard.html>`__.
