### 2023-12-30 b2187d5
Extra tweaks for py312
