+++ setup.cfg
@@ -274,7 +274,6 @@ omit =
   sunpy/cython_version*
   sunpy/*setup*
   sunpy/extern/*
-  sunpy/*/tests/*
   sunpy/version*
   sunpy/__init__*
   sunpy/data/sample.py
@@ -284,7 +283,6 @@ omit =
   */sunpy/cython_version*
   */sunpy/*setup*
   */sunpy/extern/*
-  */sunpy/*/tests/*
   */sunpy/version*
   */sunpy/__init__*
   */sunpy/data/sample.py
