### 2023-12-30 4628a78
Add text from an email from Albert - Lazy edition
