+++ .github/workflows/asv-regular.yml
@@ -27,7 +27,7 @@ jobs:
           ref: main
           path: asv_results
       - name: Set up Python 3.11
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: "3.11"
           architecture: "x64"
