### 2024-01-02 c5e1b02
[ ci skip ] Update changelog/7339.breaking.rst
Co-authored-by: David Stansby <dstansby@gmail.com>
