### 2024-01-02 7cc3593
[ skip ci ] Update docs/whatsnew/6.0.rst
Co-authored-by: David Stansby <dstansby@gmail.com>
