+++ docs/whatsnew/6.0.rst
@@ -21,8 +21,3 @@ Removal of ``sunpy.database``
 The ``sunpy.database`` module has not been actively maintained for over a year now and has a number of outstanding issues.
 It has been deprecated since sunpy 4.1, and since there have been no major objections from the communityu since then we have completely removed ``sunpy.database`` in sunpy 6.0.
 If you are interested in seeing a replacement for ``sunpy.database``, either inside ``sunpy`` or as a third-party package, please join the discussion thread at https://community.openastronomy.org/t/deprecating-sunpy-database/495.
-
-Removal of ```sunpy.map.header_helper.meta_keywords``
------------------------------------------------------
-
-``sunpy.map.header_helper.meta_keywords`` was unused.
