### 2024-01-03 01446fe
Apply suggestions from code review
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
