### 2024-01-03 7f5bc5b
sphinx_gallery_defer_figures removed
