### 2024-01-03 ae5c026
Added an example of plotting a rectangle on a map with a rotation angle relative to the axes (#7348)
* Added an example of plotting a rectangle on a map with a rotation angle relative to the axes

* reformatted the added file based on other examples, added changelog, fixed pre-commits

* fixed second plot not visible in gallery

* Applied suggestions from code review

Co-authored-by: Nabil Freij <nabil.freij@gmail.com>

* added SkyoffsetFrame link

* changed the image size to default in plot_rotated_rectangle.py and added the sphinx interlink to example file

* Apply suggestions from code revi
[…]
