### 2024-01-04 d247f04
Update docs/how_to/create_custom_map.rst
