+++ setup.cfg
@@ -235,10 +235,17 @@ filterwarnings =
     # This is raised when the VSO redirects and we do not want this to stop the CI
     ignore::sunpy.util.exceptions.SunpyConnectionWarning
 
-[pycodestyle]
-max_line_length = 110
 
 [flake8]
+ignore =
+    E225,
+    E226,
+    E501,
+    F401,
+    F403,
+    F811,
+    W503,
+    W504,
 max-line-length = 110
 exclude =
     .git,
