+++ sunpy/io/special/genx.py
@@ -19,6 +19,7 @@ class Error(Exception):
     Public ivars:
         msg -- contains the message
     """
+
     def __init__(self, msg):
         self.msg = msg
 
@@ -125,7 +126,7 @@ class Unpacker:
 
     def unpack_list(self, unpack_item):
         list = []
-        while 1:
+        while True:
             x = self.unpack_uint()
             if x == 0:
                 break
