+++ sunpy/coordinates/tests/test_metaframes.py
@@ -114,7 +114,7 @@ def test_as_base(rot_hgs):
     # Check the as_base() method
     a = rot_hgs[1].as_base()
 
-    assert isinstance(a, type(rot_hgs[1].base))  # noqa: E721
+    assert isinstance(a, type(rot_hgs[1].base))
 
     assert_longitude_allclose(a.lon, rot_hgs[1].lon)
     assert_quantity_allclose(a.lat, rot_hgs[1].lat)
@@ -170,7 +170,7 @@ def test_base_skycoord(rot_hgs):
     s = SkyCoord(1*u.deg, 2*u.deg, 3*u.AU, frame=f.HeliographicStonyhurst, obstime='2001-01-01')
     r = RotatedSunFrame(base=s)
 
-    assert isinstance(r, type(rot_hgs[1]))  # noqa: E721
+    assert isinstance(r, type(rot_hgs[1]))
     assert r.has_data
     assert not r.base.has_data
 
