+++ sunpy/coordinates/tests/test_sun.py
@@ -265,8 +265,8 @@ def test_L0_astronomical_almanac():
                  '2013-04-01': 221.44,
                  '2013-05-01': 185.30,
                  '2013-06-01': 135.30,
-                 '2013-07-01': 98.22,  # NOQA
-                 '2013-08-01': 48.03,  # NOQA
+                 '2013-07-01': 98.22,
+                 '2013-08-01': 48.03,
                  '2013-09-01': 358.28,
                  '2013-10-01': 322.22,
                  '2013-11-01': 273.31,
@@ -288,8 +288,8 @@ def test_L0_jpl_horizons():
                   '2013-04-01': 221.440599,
                   '2013-05-01': 185.306476,
                   '2013-06-01': 135.303097,
-                  '2013-07-01': 98.221806,  # NOQA
-                  '2013-08-01': 48.035951,  # NOQA
+                  '2013-07-01': 98.221806,
+                  '2013-08-01': 48.035951,
                   '2013-09-01': 358.289921,
                   '2013-10-01': 322.226009,
                   '2013-11-01': 273.315206,
