### 2024-01-04 af59818
Apply suggestions from code review
Co-authored-by: Will Barnes <will.t.barnes@gmail.com>
