+++ docs/whatsnew/6.0.rst
@@ -29,9 +29,11 @@ Arguments for :meth:`~sunpy.map.GenericMap.reproject_to`
 
 Arguments for this method have been changed to be keyword only after the target WCS argument.
 This was raising a warning since sunpy 4.1.
+
 Removal of ``sunpy.net.helioviewer.HelioViewerClient``
 ------------------------------------------------------
 
-The Helioviewer Project maintains a Python wrapper around their API that supersedes the ``sunpy.net.helioviewer`` module.
-It is called `hvpy <https://hvpy.readthedocs.io/en/latest/>`__ and their documentation has several examples of how to use it.
-It has been deprecated since sunpy 4.1.
+``sunpy.net.helioviewer`` has been deprecated since sunpy v4.1 and has now been removed.
+Users should instead use the `hvpy <https://hvpy.readthedocs.io/en/latest/>`__ package.
+This package provides a Python wrapper around the Helioviewer API and is maintained by the Helioviewer Project.
+The ``hvpy`` package supersedes the ``sunpy.net.helioviewer`` module.
