### 2024-01-04 be7f159
Backport PR #7356: Bail out of a unit test for `Helioprojective.is_visible()` if it won't be a meaningful test of tolerance
