+++ sunpy/net/tests/strategies.py
@@ -1,8 +1,7 @@
 """
 Provide a set of Hypothesis Strategies for various Fido related tests.
 """
-import sys
-import datetime
+from datetime import datetime, timezone
 
 import hypothesis.strategies as st
 import numpy as np
@@ -16,19 +15,14 @@ from astropy.time import Time
 from sunpy.net import attrs as a
 from sunpy.time import TimeRange, parse_time
 
-if sys.version_info >= (3, 12):
-    now_year = datetime.datetime.now(datetime.UTC).year
-else:
-    now_year = datetime.datetime.now().year
-
-TimesLeapsecond = sampled_from((Time('2015-06-30T23:59:60'),
+TIME_LEAP_SECONDS = sampled_from((Time('2015-06-30T23:59:60'),
                                 Time('2012-06-30T23:59:60')))
 
 
 @st.composite
 def Times(draw, max_value, min_value):
     time = one_of(datetimes(max_value=max_value, min_value=min_value),
-                  TimesLeapsecond)
+                  TIME_LEAP_SECONDS)
 
     time = Time(draw(time))
 
@@ -80,8 +74,8 @@ def online_instruments():
 
 @st.composite
 def time_attr(draw, time=Times(
-              max_value=datetime.datetime(now_year, 1, 1, 0, 0),
-              min_value=datetime.datetime(1981, 1, 1, 0, 0)),
+              max_value=datetime(datetime.now(timezone.utc).year, 1, 1, 0, 0),
+              min_value=datetime(1981, 1, 1, 0, 0)),
               delta=TimeDelta()):
     """
     Create an a.Time where it's always positive.
@@ -96,8 +90,8 @@ def time_attr(draw, time=Times(
 
 @st.composite
 def goes_time(draw, time=Times(
-              max_value=datetime.datetime(now_year, 1, 1, 0, 0),
-              min_value=datetime.datetime(1981, 1, 1, 0, 0)),
+              max_value=datetime(datetime.now(timezone.utc).year, 1, 1, 0, 0),
+              min_value=datetime(1981, 1, 1, 0, 0)),
               delta=TimeDelta()):
     """
     Create an a.Time where it's always positive.
