+++ setup.cfg
@@ -61,7 +61,7 @@ image =
   scipy>=1.7.0,!=1.10.0
 jpeg2000 =
   glymur>=0.9.1,!=0.9.5
-  lxml>=4.8.0
+  lxml>=4.8.0,!=5.0.0
 map =
   matplotlib>=3.5.0
   mpl-animators>=1.0.0
