### 2024-01-04 bad78c8
Apply suggestions from code review
Co-authored-by: David Stansby <dstansby@gmail.com>
