+++ changelog/7342.trivial.rst
@@ -1,1 +1,1 @@
-`~sunpy.net.jsoc.attrs.Notify` checks for a valid input value.
+`~sunpy.net.jsoc.attrs.Notify` checks that a valid email address has been given as a value.
