### 2024-01-04 083cdfa
Update the precommit to have a global exclude
