### 2024-01-04 dcc3ebe
Update sunpy/net/jsoc/jsoc.py
Co-authored-by: Will Barnes <will.t.barnes@gmail.com>
