### 2024-01-04 7dcca41
Apply suggestions from code review
Co-authored-by: Will Barnes <will.t.barnes@gmail.com>
