+++ sunpy/net/vso/tests/test_vso.py
@@ -24,7 +24,7 @@ from sunpy.net.vso.vso import (
 )
 from sunpy.tests.mocks import MockObject
 from sunpy.time import parse_time
-from sunpy.util.exceptions import SunpyConnectionWarning, SunpyUserWarning
+from sunpy.util.exceptions import SunpyConnectionWarning, SunpyDeprecationWarning, SunpyUserWarning
 
 
 class MockQRRecord:
@@ -122,10 +122,11 @@ def test_path(client, tmpdir):
     Test that '{file}' is automatically appended to the end of a custom path if
     it is not specified.
     """
-    qr = client.search(
-        core_attrs.Time('2020-06-07 06:33', '2020-06-07 06:33:13'),
-        core_attrs.Instrument('aia'), core_attrs.Wavelength(171 * u.AA),
-        response_format="table")
+    with pytest.warns(SunpyDeprecationWarning, match="response_format"):
+        qr = client.search(
+            core_attrs.Time('2020-06-07 06:33', '2020-06-07 06:33:13'),
+            core_attrs.Instrument('aia'), core_attrs.Wavelength(171 * u.AA),
+            response_format="table")
     tmp_dir = tmpdir / "{file}"
     files = client.fetch(qr, path=tmp_dir)
 
@@ -156,7 +157,8 @@ def test_no_download(client):
     stereo = (core_attrs.Detector('STEREO_B') &
               core_attrs.Instrument('EUVI') &
               core_attrs.Time('1900-01-01', '1900-01-01T00:10:00'))
-    qr = client.search(stereo, response_format="table")
+    with pytest.warns(SunpyDeprecationWarning, match="response_format"):
+        qr = client.search(stereo, response_format="table")
     downloader = MockDownloader()
     res = client.fetch(qr, wait=False, downloader=downloader)
     assert downloader.download_called is False
@@ -166,8 +168,7 @@ def test_no_download(client):
 def test_non_str_instrument():
     # Sanity Check
     assert isinstance(core_attrs.Instrument("lyra"), core_attrs.Instrument)
-
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match="Instrument names must be strings"):
         core_attrs.Instrument(1234)
 
 
@@ -250,8 +251,9 @@ def test_vso_hmi(client, tmpdir):
     """
     This is a regression test for https://github.com/sunpy/sunpy/issues/2284
     """
-    res = client.search(core_attrs.Time('2020-01-02 23:52:00', '2020-01-02 23:54:00'),
-                        core_attrs.Instrument('HMI') | core_attrs.Instrument('AIA'), response_format="table")
+    with pytest.warns(SunpyDeprecationWarning, match="response_format"):
+        res = client.search(core_attrs.Time('2020-01-02 23:52:00', '2020-01-02 23:54:00'),
+                            core_attrs.Instrument('HMI') | core_attrs.Instrument('AIA'), response_format="table")
 
     dr = client.make_getdatarequest(res)
 
@@ -337,27 +339,28 @@ def test_VSOClient(mocker):
     Unable to find any valid VSO mirror? Raise ConnectionError
     """
     mocker.patch('sunpy.net.vso.vso.get_online_vso_url', return_value=None)
-    with pytest.raises(ConnectionError):
+    with pytest.raises(ConnectionError, match="No online VSO mirrors could be found."):
         VSOClient()
 
 
 def test_build_client(mocker):
     mocker.patch('sunpy.net.vso.vso.check_connection', return_value=None)
-    with pytest.raises(ConnectionError):
+    with pytest.raises(ConnectionError, match="Can't connect to url http://notathing.com/"):
         build_client(url="http://notathing.com/", port_name="spam")
 
 
 def test_build_client_params():
[…]
