+++ sunpy/net/hek2vso/hek2vso.py
@@ -1,10 +1,6 @@
-# Author:   Michael Malocha <mjm159@humboldt.edu>
-# Last Edit:  August 10th, 2013
-#
-# This module was developed with funding from the GSOC 2013 summer of code
 """
-This module translates the results of a HEK query into a VSO query
-and returns the results from the VSO query to the user.
+This module translates the results of a HEK query into a VSO query and
+returns the results from the VSO query to the user.
 """
 
 import sys
@@ -16,9 +12,6 @@ from astropy import units
 from sunpy.net import attrs as a
 from sunpy.net import hek, vso
 from sunpy.net.hek import HEKTable
-
-__author__ = 'Michael Malocha'
-__version__ = 'Aug 10th, 2013'
 
 __all__ = ['translate_results_to_query', 'vso_attribute_parse', 'H2VClient']
 
@@ -140,8 +133,10 @@ class H2VClient:
         Examples
         --------
         >>> from sunpy.net import attrs as a, hek, hek2vso
-        >>> h2v = hek2vso.H2VClient()  # doctest: +REMOTE_DATA
-        >>> q = h2v.full_query((a.Time('2011/08/09 07:23:56', '2011/08/09 12:40:29'), a.hek.EventType('FL')))  # doctest: +REMOTE_DATA
+        >>> hek2vso_client = hek2vso.H2VClient()  # doctest: +REMOTE_DATA
+        >>> query = hek2vso_client.full_query((a.Time('2011/08/09 07:00:00', '2011/08/09 07:15:00'), a.hek.EventType('FL')))  # doctest: +REMOTE_DATA
+        >>> len(query)  # doctest: +REMOTE_DATA
+        7
         """
         self._quick_clean()
         if progress:
@@ -150,9 +145,9 @@ class H2VClient:
         self.hek_results = self.hek_client.search(*client_query)
         self._quick_clean()
         return self.translate_and_query(self.hek_results,
-                                        limit=limit, progress=progress)
+                                        limit=limit)
 
-    def translate_and_query(self, hek_results, limit=None, progress=False, vso_response_format="table"):
+    def translate_and_query(self, hek_results, limit=None):
         """
         Translates HEK results, makes a VSO query, then returns the results.
 
@@ -183,7 +178,7 @@ class H2VClient:
         vso_query = translate_results_to_query(hek_results)
 
         for query in tqdm(vso_query, unit="records"):
-            temp = self.vso_client.search(*query, response_format=vso_response_format)
+            temp = self.vso_client.search(*query)
             self.vso_results.append(temp)
             self.num_of_records += len(temp)
             if limit is not None:
