### 2024-01-05 8e711a3
Apply suggestions from code review
Co-authored-by: David Stansby <dstansby@gmail.com>
Co-authored-by: Will Barnes <will.t.barnes@gmail.com>
