+++ README.rst
@@ -2,52 +2,110 @@
 ``sunpy``
 *********
 
-|Latest PyPi Version|
-|Latest Conda Version|
-|DOI|
-|repostatus|
-|python|
-|ci|
-|codecov|
-|Docs|
-|matrix|
-|Powered by NumFOCUS|
+.. raw:: html
 
-.. |Latest PyPi Version| image:: https://img.shields.io/pypi/v/sunpy.svg
-    :target: https://pypi.python.org/pypi/sunpy/
-    :alt: Latest Pypi Version
-.. |Latest Conda Version| image:: https://anaconda.org/conda-forge/sunpy/badges/version.svg
-    :target: https://anaconda.org/conda-forge/sunpy
-    :alt: Latest Conda Version
-.. |DOI| image:: https://zenodo.org/badge/2165383.svg
-    :target: https://zenodo.org/badge/latestdoi/2165383
-    :alt: Zenodo - Latest DOI
-.. |matrix| image:: https://img.shields.io/matrix/sunpy:openastronomy.org.svg?colorB=%23FE7900&label=Chat&logo=matrix&server_fqdn=matrix.org
-    :target: https://app.element.io/#/room/#sunpy:openastronomy.org
-    :alt: Matrix Chat Room
-.. |codecov| image:: https://codecov.io/gh/sunpy/sunpy/branch/main/graph/badge.svg
-    :target: https://codecov.io/gh/sunpy/sunpy
-    :alt: Code Coverage
-.. |Powered by NumFOCUS| image:: https://img.shields.io/badge/powered%20by-NumFOCUS-orange.svg?style=flat&colorA=E1523D&colorB=007D8A
-    :target: https://numfocus.org
-    :alt: Powered by NumFOCUS
-.. |CI| image:: https://github.com/sunpy/sunpy/actions/workflows/ci.yml/badge.svg?branch=main
-    :target: https://github.com/sunpy/sunpy/actions/workflows/ci.yml
-    :alt: Continuous Integration Status
-.. |Docs| image:: https://readthedocs.org/projects/sunpy/badge/?version=stable
-    :target: https://docs.sunpy.org/en/stable/?badge=stable
-    :alt: Stable Documentation
-.. |repostatus| image:: https://www.repostatus.org/badges/latest/active.svg
-    :target: https://www.repostatus.org/#active
-    :alt: Project Status: Active - The project has reached a stable, usable state and is being actively developed.
-.. |python| image:: https://img.shields.io/pypi/pyversions/sunpy
-    :alt: PyPI - Python Version
+    <img
+    src="https://raw.githubusercontent.com/sunpy/sunpy-logo/master/sunpy_logo_landscape.svg"
+    alt="SunPy Logo"
+    align="center"
+    width="200px" />
+    <table>
+    <thead>
+        <tr>
+        <th> Release </th>
+        <th> Development </th>
+        <th> Community </th>
+        </tr>
+    </thead>
+    <tbody>
+        <tr>
+        <td>
+            <a href="https://pypi.python.org/pypi/sunpy/"
+            ><img src="https://img.shields.io/pypi/v/sunpy.svg" alt="Latest PyPi Version"
+            /></a>
+        </td>
+        <td>
+            <a href="https://pypi.python.org/pypi/sunpy/"
+            ><img src="https://img.shields.io/pypi/pyversions/sunpy" alt="Python Versions"
+            /></a>
+        </td>
+        <td>
+            <a href="https://app.element.io/#/room/#sunpy:openastronomy.org"
+            ><img
+                src="https://img.shields.io/matrix/sunpy:openastronomy.org.svg?colorB=%23FE7900&label=Chat&logo=matrix&server_fqdn=matrix.org"
[…]
