### 2024-01-05 7a9d395
Merge pull request #7325 from nabobalis/readme
Address review comments from the pyopensci reviewers
