+++ setup.cfg
@@ -96,6 +96,7 @@ all =
   %(timeseries)s
   %(visualization)s
 tests =
+  %(all)s
   hvpy>=1.0.1
   hypothesis>=6.0.0  # Included in pytest-astropy. 6.0 is the first version to support disabling function-scoped fixture warning
   jplephem  # For some coordinates tests
@@ -107,6 +108,7 @@ tests =
   pytest-xdist>=2.0
   pytest>=6.0
 docs =
+  %(all)s
   astroquery
   hvpy>=1.0.1
   jplephem
@@ -152,7 +154,7 @@ testpaths = "sunpy" "docs"
 norecursedirs = ".tox" "build" "docs[\/]_build" "docs[\/]generated" "*.egg-info" "examples" "sunpy[/\]_dev" ".jupyter" ".history" "tools" "sunpy[\/]extern" "benchmarks"
 doctest_plus = enabled
 doctest_optionflags = NORMALIZE_WHITESPACE FLOAT_CMP ELLIPSIS
-addopts = --dist no --arraydiff --doctest-rst --doctest-ignore-import-errors -p no:unraisableexception -p no:threadexception
+addopts = --dist no --arraydiff --doctest-rst --doctest-ignore-import-errors -p no:unraisableexception -p no:threadexception -m "not mpl_image_compare"
 asdf_schema_tests_enabled = true
 asdf_schema_root = sunpy/io/special/asdf/resources/
 mpl-results-path = figure_test_images
