### 2024-01-05 e1302c9
Apply suggestions from code review
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
