### 2024-01-05 58bb2a0
Apply suggestions from code review
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
