+++ sunpy/map/tests/test_plotting.py
@@ -228,7 +228,7 @@ def test_draw_contours_different_wcs(aia171_test_map):
 
 
 @figure_test
-def test_draw_contours_aia_full(aia171_test_map):
+def test_draw_contours_aia_fill(aia171_test_map):
     aia171_test_map.plot()
     aia171_test_map.draw_contours(u.Quantity(np.arange(1, 100, 10), 'percent'), fill=True)
 
