### 2024-01-05 1a72fe2
Apply suggestions from code review
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
