### 2024-01-06 39ced47
Add checks for installed extras
tweak warning msg

prevent warning on no errors
