+++ sunpy/net/__init__.py
@@ -1,5 +1,9 @@
 
 # Import and register the clients but we do not want them in the namespace, we import them as _
+# check if user has installed the net extras
+from sunpy.util.sysinfo import warn_missing_deps
+warn_missing_deps('net')
+
 from sunpy.net import base_client as _
 from sunpy.net import cdaweb as _
 from sunpy.net import dataretriever as _
