+++ sunpy/image/__init__.py
@@ -1,0 +1,3 @@
+# check if user has installed the image extras
+from sunpy.util.sysinfo import warn_missing_deps
+warn_missing_deps('image')
