+++ sunpy/net/__init__.py
@@ -2,6 +2,7 @@
 # Import and register the clients but we do not want them in the namespace, we import them as _
 # check if user has installed the net extras
 from sunpy.util.sysinfo import warn_missing_deps
+
 warn_missing_deps('net')
 
 from sunpy.net import base_client as _
