+++ sunpy/timeseries/__init__.py
@@ -1,5 +1,6 @@
 # check if user has installed the timeseries extras
 from sunpy.util.sysinfo import warn_missing_deps
+
 warn_missing_deps('timeseries')
 
 from sunpy.timeseries.metadata import TimeSeriesMetaData
