+++ sunpy/visualization/__init__.py
@@ -1,5 +1,6 @@
 # check if user has installed the visualisation extras
 from sunpy.util.sysinfo import warn_missing_deps
+
 warn_missing_deps('visualization')
 
 from sunpy.visualization.visualization import *
