+++ sunpy/image/__init__.py
@@ -1,3 +1,4 @@
 # check if user has installed the image extras
 from sunpy.util.sysinfo import warn_missing_deps
+
 warn_missing_deps('image')
