### 2024-01-06 4d3cec1
moved importwarning filter inside missing deps check
swapped to warn user

capitilsation of comments, block warn function from export
