+++ sunpy/visualization/__init__.py
@@ -1,6 +1,5 @@
-# check if user has installed the visualisation extras
-from sunpy.util.sysinfo import warn_missing_deps
-
-warn_missing_deps('visualization')
+# Check if user has installed the visualisation extras
+from sunpy.util.sysinfo import warn_missing_deps as _warn_missing_deps
+_warn_missing_deps('visualization')
 
 from sunpy.visualization.visualization import *
