+++ sunpy/timeseries/__init__.py
@@ -1,7 +1,6 @@
-# check if user has installed the timeseries extras
-from sunpy.util.sysinfo import warn_missing_deps
-
-warn_missing_deps('timeseries')
+# Check if user has installed the timeseries extras
+from sunpy.util.sysinfo import warn_missing_deps as _warn_missing_deps
+_warn_missing_deps('timeseries')
 
 from sunpy.timeseries.metadata import TimeSeriesMetaData
 from sunpy.timeseries.sources import *
