### 2024-01-07 2a023a0
Adds warning in reproject_to() method
Adds changelog, a unit test and changes to reproject_to() method.
