### 2024-01-08 d8e47c9
Norm set for data which is not uint8
