### 2024-01-10 818b187
fixes two notes section in the docs
