### 2024-01-10 c36512b
Merge pull request #7351 from nabobalis/py312
Add python 3.12 tests

(cherry picked from commit 7961e296893d52d2152c1f8ac0ba8f481813e28a)
