+++ setup.cfg
@@ -30,6 +30,7 @@ classifiers =
   Programming Language :: Python :: 3.9
   Programming Language :: Python :: 3.10
   Programming Language :: Python :: 3.11
+  Programming Language :: Python :: 3.12
   Topic :: Scientific/Engineering :: Physics
 
 [options]
@@ -61,7 +62,7 @@ image =
   scipy>=1.7.0,!=1.10.0
 jpeg2000 =
   glymur>=0.8.18,!=0.9.0,!=0.9.5
-  lxml>=4.8.0
+  lxml>=4.8.0,!=5.0.0
 map =
   matplotlib>=3.5.0
   mpl-animators>=1.0.0
@@ -236,6 +237,8 @@ filterwarnings =
     ignore:.*module is deprecated, as it was designed for internal use
     # This is raised when the VSO redirects and we do not want this to stop the CI
     ignore::sunpy.util.exceptions.SunpyConnectionWarning
+    # Can be removed when https://github.com/dateutil/dateutil/issues/1314 is resolved
+    ignore:datetime.datetime.utcfromtimestamp():DeprecationWarning
 
 [pycodestyle]
 max_line_length = 110
@@ -273,7 +276,6 @@ omit =
   sunpy/cython_version*
   sunpy/*setup*
   sunpy/extern/*
-  sunpy/*/tests/*
   sunpy/version*
   sunpy/__init__*
   sunpy/data/sample.py
@@ -283,7 +285,6 @@ omit =
   */sunpy/cython_version*
   */sunpy/*setup*
   */sunpy/extern/*
-  */sunpy/*/tests/*
   */sunpy/version*
   */sunpy/__init__*
   */sunpy/data/sample.py
