### 2024-01-10 63f13a3
Backport PR #7351: Add python 3.12 tests
