### 2024-01-10 5fa723f
Update setup.cfg
