+++ setup.cfg
@@ -240,14 +240,14 @@ filterwarnings =
 
 [flake8]
 ignore =
-    E225,
-    E226,
-    E501,
-    F401,
-    F403,
-    F811,
-    W503,
-    W504,
+    E225, # missing-whitespace-around-operator
+    E226, # missing-whitespace-around-arithmetic-operator
+    E501, # line-too-long
+    F401, # unused-import
+    F403, # undefined-local-with-import-star
+    F811, # redefined-while-unused
+    W503, # Line break occurred before a binary operator
+    W504, # Line break occurred after a binary operator
 max-line-length = 110
 exclude =
     .git,
