### 2024-01-10 9c4c953
Update docs/how_to/fix_map_metadata.rst
Co-authored-by: Will Barnes <will.t.barnes@gmail.com>
