### 2024-01-10 d47ab9f
Backport PR #7343: Ran the autopep8 linter for pyopensci review
