+++ sunpy/image/__init__.py
@@ -1,3 +1,4 @@
 # Check if user has installed the image extras
-from sunpy.util.sysinfo import warn_missing_deps as _warn_missing_deps
+from sunpy.util.sysinfo import _warn_missing_deps
+
 _warn_missing_deps('image')
