+++ sunpy/map/__init__.py
@@ -4,7 +4,7 @@ SunPy Map
 isort:skip_file
 """
 # Check if user has installed the map extras
-from sunpy.util.sysinfo import warn_missing_deps as _warn_missing_deps
+from sunpy.util.sysinfo import _warn_missing_deps
 _warn_missing_deps('map')
 
 from ._units import maxwell
