+++ sunpy/timeseries/__init__.py
@@ -1,5 +1,6 @@
 # Check if user has installed the timeseries extras
-from sunpy.util.sysinfo import warn_missing_deps as _warn_missing_deps
+from sunpy.util.sysinfo import _warn_missing_deps
+
 _warn_missing_deps('timeseries')
 
 from sunpy.timeseries.metadata import TimeSeriesMetaData
