### 2024-01-11 afe9a0d
Update sunpy/net/__init__.py
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
