### 2024-01-11 69be60c
Backport PR #7378: add replacement to mapbase for unit: counts / pixel -> ct/pix
