### 2024-01-11 b789d17
Backport PR #7378: add replacement to mapbase for unit: counts / pixel -> ct/pix
