+++ setup.cfg
@@ -152,7 +152,7 @@ use_extension_helpers = true
 
 [tool:pytest]
 testpaths = "sunpy" "docs"
-norecursedirs = ".tox" "build" "docs[\/]_build" "docs[\/]generated" "*.egg-info" "examples" "sunpy[/\]_dev" ".jupyter" ".history" "tools" "sunpy[\/]extern" "benchmarks"
+norecursedirs = ".tox" "build" "docs[\/]_build" "docs[\/]generated" "*.egg-info" "examples" "sunpy[/\]_dev" ".history" "tools" "sunpy[\/]extern" "benchmarks"
 doctest_plus = enabled
 doctest_optionflags = NORMALIZE_WHITESPACE FLOAT_CMP ELLIPSIS
 addopts = --dist no --arraydiff --doctest-rst --doctest-ignore-import-errors -p no:unraisableexception -p no:threadexception -m "not mpl_image_compare"
@@ -240,10 +240,17 @@ filterwarnings =
     # Can be removed when https://github.com/dateutil/dateutil/issues/1314 is resolved
     ignore:datetime.datetime.utcfromtimestamp():DeprecationWarning
 
-[pycodestyle]
-max_line_length = 110
 
 [flake8]
+ignore =
+    E225, # missing-whitespace-around-operator
+    E226, # missing-whitespace-around-arithmetic-operator
+    E501, # line-too-long
+    F401, # unused-import
+    F403, # undefined-local-with-import-star
+    F811, # redefined-while-unused
+    W503, # Line break occurred before a binary operator
+    W504, # Line break occurred after a binary operator
 max-line-length = 110
 exclude =
     .git,
