### 2024-01-11 a3679a7
Fix manual type checks
