### 2024-01-11 68ce16a
Removal maxwell unit and bump astropy to 5.1
