+++ setup.cfg
@@ -41,8 +41,7 @@ include_package_data = True
 setup_requires =
   setuptools_scm
 install_requires =
-  # Exclude versions of astropy which are incompatible with matplotlib 3.7
-  astropy>=5.0.6,!=5.1.0
+  astropy>=5.1.1
   numpy>=1.21.0
   packaging>=19.0
   parfive[ftp]>=2.0.0
