### 2024-01-12 3dd7013
Update changelog/7364.deprecation.rst
Co-authored-by: David Stansby <dstansby@gmail.com>
