+++ setup.cfg
@@ -241,6 +241,9 @@ filterwarnings =
     ignore:datetime.datetime.utcfromtimestamp():DeprecationWarning
     # https://github.com/pandas-dev/pandas/issues/54466
     ignore:\nPyarrow will become a required dependency of pandas in the next major release of pandas:DeprecationWarning
+    # Pending removal from sunpy 7.0
+    ignore:The QueryResponse class is deprecated
+
 
 [flake8]
 ignore =
