+++ changelog/7307.trivial.rst
@@ -2,10 +2,22 @@
 This has been removed in `drms` 0.7.0, which had breaking API changes within this release.
 As a result, we increased the minimum required version of `drms` to 0.7.0.
 
-This specifically refers to the following information that was printed to the console:
+This specifically refers to the following information that was printed to the console by default:
 
 ``"Export request pending. [id=X, status=X]"``
 ``"Waiting for X seconds..."``
 ``"Request not found on server, X retries left."``
 
 These were handled by `drms` and are now logging messages.
+
+If you want to silence these messages, you can set the logging level to ``WARNING`` or higher.
+
+.. code-block:: python
+
+    import logging
+    drms_logger = logging.getLogger("drms")
+    drms_logger.setLevel(logging.WARNING)
+
+    from sunpy.net import fido, attrs
+
+Note, you have to do it before you import ``fido``.
