+++ setup.cfg
@@ -248,14 +248,22 @@ filterwarnings =
 
 [flake8]
 ignore =
-    E225, # missing-whitespace-around-operator
-    E226, # missing-whitespace-around-arithmetic-operator
-    E501, # line-too-long
-    F401, # unused-import
-    F403, # undefined-local-with-import-star
-    F811, # redefined-while-unused
-    W503, # Line break occurred before a binary operator
-    W504, # Line break occurred after a binary operator
+    # missing-whitespace-around-operator
+    E225,
+    # missing-whitespace-around-arithmetic-operator
+    E226,
+    # line-too-long
+    E501,
+    # unused-import
+    F401,
+    # undefined-local-with-import-star
+    F403,
+    # redefined-while-unused
+    F811,
+    # Line break occurred before a binary operator
+    W503,
+    # Line break occurred after a binary operator
+    W504,
 max-line-length = 110
 exclude =
     .git,
