### 2024-01-15 #147 @nabobalis
Hello everyone,

Sorry again for the slowness from our side:

The following PRs were merged in:

https://github.com/sunpy/sunpy/pull/7325
https://github.com/sunpy/sunpy/pull/7351
https://github.com/sunpy/sunpy/pull/7343

This should address ~85% of the review comments.
These form the new releases 5.0.3 and 5.1.1 which are out on pypi and conda-forge.

Hopefully these will be good enough but I am happy to address any issues that I missed.

I plan to follow up in a few weeks with a more comprehensive rewrite of the developer's guide to tie in with the installation guide suggestions from this review.

You will also notice we added several flake8 ignore rules due to failing to achieve consensus over a full style and that will require a format that will allow us to not put spaces around specific operators that deal with astropy units.

The readme table in the release version is actually a rst table where as the one in main is a raw::html, that will change in main with another PR (https://github.com/sunpy/sunpy/pull/7384), it also updates the packaging of sunpy to use nothing but pyproject.toml.

There is a different PR that needs another review which will raise a message if you import a submodule without the optional dependencies installed (https://github.com/sunpy/sunpy/pull/7369)

I plan to also submit a pr removing the `import attrs as a` style we have been using.
