### 2024-01-15 43f6ef8
Increase min versions for a 6.0 release in ~ may 2024
