+++ changelog/7383.trivial.rst
@@ -1,0 +1,1 @@
+Increased minimum version of `astropy` to 5.3.0.
