+++ setup.cfg
@@ -41,8 +41,8 @@ include_package_data = True
 setup_requires =
   setuptools_scm
 install_requires =
-  astropy>=5.1.1
-  numpy>=1.21.0
+  astropy>=5.3.0
+  numpy>=1.21.0,<2
   packaging>=19.0
   parfive[ftp]>=2.0.0
   requests
@@ -50,39 +50,39 @@ install_requires =
 [options.extras_require]
 # The list of available extras is also in the installation docs, if you add or remove one please edit it there as well.
 asdf =
-  asdf-astropy>=0.1.1
-  asdf>=2.8.0
+  asdf-astropy>=0.2.0
+  asdf>=2.11.0
 dask =
-  dask[array]>=2021.4.0
+  dask[array]>=2022.5.2
 image =
-  scikit-image>=0.18.0
+  scikit-image>=0.19.0
   # !=1.10.0 due to https://github.com/scipy/scipy/issues/17718
-  scipy>=1.7.0,!=1.10.0
+  scipy>=1.8.0,!=1.10.0
 jpeg2000 =
   glymur>=0.9.1,!=0.9.5
-  lxml>=4.8.0,!=5.0.0
+  lxml>=4.9.0,!=5.0.0
 map =
   matplotlib>=3.5.0
   mpl-animators>=1.0.0
-  reproject
+  reproject>=0.9.0
   # !=1.10.0 due to https://github.com/scipy/scipy/issues/17718
-  scipy>=1.7.0,!=1.10.0
+  scipy>=1.8.0,!=1.10.0
 net =
-  beautifulsoup4>=4.8.0
+  beautifulsoup4>=4.11.0
   drms>=0.7.1
   python-dateutil>=2.8.0
-  tqdm>=4.32.1
-  zeep>=3.4.0
+  tqdm>=4.64.0
+  zeep>=4.1.0
 spice =
-  spiceypy
+  spiceypy>=5.0.0
 timeseries =
-  cdflib>=0.3.20,!=0.4.0,!=1.0.0
-  h5netcdf>=0.11
+  cdflib>=0.4.4,!=1.0.0
+  h5netcdf>=1.0.0
   # While a not direct dependency
   # We need to raise this to ensure the goes netcdf files open.
-  h5py>=3.1.0
+  h5py>=3.7.0
   matplotlib>=3.5.0
-  pandas>=1.2.0
+  pandas>=1.4.0
 visualization =
   matplotlib>=3.5.0
   mpl-animators>=1.0.0
@@ -97,33 +97,32 @@ all =
   %(visualization)s
 tests =
   %(all)s
-  hvpy>=1.0.1
-  hypothesis>=6.0.0  # Included in pytest-astropy. 6.0 is the first version to support disabling function-scoped fixture warning
+  hvpy
   jplephem  # For some coordinates tests
   opencv-python
-  pytest-astropy>=0.8  # 0.8 is the first release to include filter-subpackage
-  pytest-doctestplus>=0.5 # We require the newest version of doctest plus to use +IGNORE_WARNINGS
+  pytest-astropy
+  pytest-doctestplus
   pytest-mock
-  pytest-mpl>=0.12 # First version to support our figure tests
-  pytest-xdist>=2.0
-  pytest>=6.0
+  pytest-mpl
+  pytest-xdist
+  pytest
   %(spice)s
 docs =
   %(all)s
-  hvpy>=1.0.1
+  hvpy
   packaging
   ruamel.yaml
   sphinx
   sphinx-automodapi
-  # Remove next line when fixed in towncrier; see https://github.com/twisted/towncrier/issues/528
+  # Remove next line when fixed in sphinx-changelog; see https://github.com/OpenAstronomy/sphinx-changelog/issues/22
   importlib-resources<6
   sphinx-changelog
[…]
