### 2024-01-15 a7bd463
Increase more versions plus update oldestdeps and figure tests
