+++ changelog/7383.trivial.rst
@@ -1,1 +1,19 @@
-Increased minimum version of `astropy` to 5.3.0.
+Increased minimum versions for:
+
+- astropy >= 5.3.0
+- asdf-astropy >= 0.2.0
+- asdf >= 2.11.0
+- dask >= 2022.5.2
+- scikit-image >= 0.19.0
+- scipy >= 1.8.0
+- lxml >= 4.9.0
+- reproject>=0.9.0
+- beautifulsoup4 >= 4.11.0
+- tqdm >= 4.64.0
+- zeep>=4.1.0
+- spiceypy>=5.0.0
+- cdflib >= 0.4.4
+- h5netcdf > =1.0.0
+- h5py >= 3.7.0
+- pandas >= 1.4.0
+- opencv-python >= 4.6.0.66
