+++ setup.cfg
@@ -67,6 +67,8 @@ map =
   reproject>=0.9.0
   # !=1.10.0 due to https://github.com/scipy/scipy/issues/17718
   scipy>=1.8.0,!=1.10.0
+opencv =
+  opencv-python>=4.6.0.66
 net =
   beautifulsoup4>=4.11.0
   drms>=0.7.1
@@ -99,13 +101,13 @@ tests =
   %(all)s
   hvpy
   jplephem  # For some coordinates tests
-  opencv-python
+  %(opencv)s
   pytest-astropy
   pytest-doctestplus
   pytest-mock
   pytest-mpl
   pytest-xdist
-  pytest
+  pytest >= 7.1.0
   %(spice)s
 docs =
   %(all)s
@@ -125,11 +127,11 @@ docs =
   sunpy-sphinx-theme
   sphinx-hoverxref
 docs-gallery =
+  %(docs)s
   astroquery
   jplephem
   mplcairo
-  opencv-python
-  reproject
+  %(opencv)s
 dev =
   %(all)s
   %(tests)s
