### 2024-01-16 dd7be9d
Downgrade to astropy 5.2.0
