+++ sunpy/util/tests/test_sysinfo.py
@@ -9,45 +9,46 @@ from sunpy.util.sysinfo import (
 )
 
 EXTRA_DEPS = [
-    'asdf',
     'asdf-astropy',
+    'asdf',
     'astropy',
-    'numpy',
-    'parfive',
-    'packaging',
+    'beautifulsoup4',
+    'cdflib',
     'dask',
-    'scikit-image',
-    'scipy',
+    'drms',
+    'h5netcdf',
+    'h5py',
     'lxml',
     'matplotlib',
     'mpl-animators',
-    'reproject',
-    'beautifulsoup4',
-    'drms',
+    'numpy',
+    'packaging',
+    'pandas',
+    'parfive',
     'python-dateutil',
+    'reproject',
+    'scikit-image',
+    'scipy',
     'zeep',
-    'cdflib',
-    'h5netcdf',
-    'h5py',
-    'pandas'
 ]
 
 EXTRA_ALL_GROUPS = [
     'all',
     'asdf',
-    'required',
     'dask',
     'dev',
-    'docs',
     'docs-gallery',
+    'docs',
     'image',
     'jpeg2000',
     'map',
     'net',
+    'opencv',
+    'required',
     'spice',
     'tests',
     'timeseries',
-    'visualization'
+    'visualization',
 ]
 
 def test_find_dependencies():
