+++ changelog/7383.trivial.rst
@@ -1,6 +1,6 @@
 Increased minimum versions for:
 
-- astropy >= 5.3.0
+- astropy >= 5.2.0
 - asdf-astropy >= 0.2.0
 - asdf >= 2.11.0
 - dask >= 2022.5.2
