+++ setup.cfg
@@ -41,7 +41,7 @@ include_package_data = True
 setup_requires =
   setuptools_scm
 install_requires =
-  astropy>=5.3.0
+  astropy>=5.2.0
   numpy>=1.21.0,<2
   packaging>=19.0
   parfive[ftp]>=2.0.0
