+++ tox.ini
@@ -43,7 +43,7 @@ deps =
     # oldestdeps we pin against to ensure correct versions
     oldestdeps: asdf-astropy<0.3.0
     oldestdeps: asdf<2.12.0
-    oldestdeps: astropy<6.0.0
+    oldestdeps: astropy<5.3.0
     oldestdeps: beautifulsoup4<4.12.0
     oldestdeps: cdflib<0.4.5
     oldestdeps: dask[array]<2022.6.0
