### 2024-01-18 #147 @Septaris
Hello @nabobalis,

You've done a great job 🎉, and indeed, you've addressed the main issues. Regarding the ignore rules, as long as you know what you're doing, it's fine. I hope you'll move towards a fully automated solution for code formatting to save you from having to manually format the code and/or add flake8 ignore rules.

Good luck for the rewrite of the developer's guide! 😃

@cmarmo I think we're ready to conclude the review.
