### 2024-01-18 #147 @cmarmo
Hello everybody,

 🎉 it is my pleasure to announce that SunPy has been approved by pyOpenSci! Thank you @nabobalis  for submitting SunPy and many thanks to @nutjob4life and @Septaris for reviewing this package! 😸

If you all could find some time to fill out our [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7) this will help us a lot! Thanks!

Also, let us know if you are interested in joining the pyOpenSci Slack (if you are not already there).

- [ ] Add the badge for pyOpenSci peer-review to the README.md of SunPy. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`. .... sorry the table would become even more challenging...

---
