### 2024-01-18 #147 @nabobalis
Thank you again @cmarmo, @nutjob4life and @Septaris.

I have added the badge to this PR: https://github.com/sunpy/sunpy/pull/7384/files?short_path=7b3ed02#diff-7b3ed02bc73dc06b7db906cf97aa91dec2b2eb21f2d92bc5caa761df5bbc168f

I put it under the community section in a new row. I think maybe a 4th column would be better as I also want to add a link to our joss paper and our latest published paper on sunpy.

I have updated the release branches separately.

> @nabobalis , when you and the SunPy community think it would be appropriate, please let us know how to communicate about this review (blog post, Discourse announcement, ....), thanks!

I will ask our social manager but I would be leaning towards adding a blogpost on sunpy.org and a discourse post on our openastro instance.

Would pyopensci post an announcement on their own blog or discourse?
