### 2024-01-20 #147 @cmarmo
> Would pyopensci post an announcement on their own blog or discourse?

The editor generally announce the successful review on [discourse](https://pyopensci.discourse.group/c/packages/8): I would be happy to cross-link any announcement on your side on that same occasion.
