### 2024-01-22 ecd4cf1
Update docs/topic_guide/fits_keywords.rst
Co-authored-by: Andrew Robbertz <24920994+Alrobbertz@users.noreply.github.com>
