### 2024-01-22 f87451c
Add python 310 bump that comes from pyproject pr
