+++ docs/whatsnew/6.0.rst
@@ -19,6 +19,8 @@ On this page, you can read about some of the big changes in this release.
 Updates to minimum dependencies
 ===============================
 
+The minimum required version of Python has been updated to 3.10.
+
 The minimum required versions of core dependencies have been updated:
 
 - astropy >= 5.2.0
