+++ setup.cfg
@@ -42,7 +42,7 @@ setup_requires =
   setuptools_scm
 install_requires =
   astropy>=5.2.0
-  numpy>=1.21.0,<2
+  numpy>=1.21.0
   packaging>=19.0
   parfive[ftp]>=2.0.0
   requests
