### 2024-01-22 9b8e53c
Backport PR #7391: ignore filter for pandas
