+++ sunpy/timeseries/timeseriesbase.py
@@ -866,7 +866,7 @@ class GenericTimeSeries:
         Parameters
         ----------
         **kwargs : `dict`
-            All keyword arguments are passed to `pandas.DataFrame.to_numpy`.
+            All keyword arguments are passed to ``pandas.DataFrame.to_numpy``.
 
         Returns
         -------
