+++ setup.cfg
@@ -239,7 +239,8 @@ filterwarnings =
     ignore::sunpy.util.exceptions.SunpyConnectionWarning
     # Can be removed when https://github.com/dateutil/dateutil/issues/1314 is resolved
     ignore:datetime.datetime.utcfromtimestamp():DeprecationWarning
-
+    # https://github.com/pandas-dev/pandas/issues/54466
+    ignore:\nPyarrow will become a required dependency of pandas in the next major release of pandas:DeprecationWarning
 
 [flake8]
 ignore =
