### 2024-01-22 9d97f7a
Update changelog/7307.trivial.rst
Co-authored-by: David Stansby <dstansby@gmail.com>
