### 2024-01-23 58c387b
notes heading reappearing fix
