### 2024-01-24 3cd2e7d
Removes variables `rsun_map` and `rsun_target` from reproject_to()
