+++ sunpy/map/mapbase.py
@@ -2740,11 +2740,9 @@ class GenericMap(NDData):
                             plot_settings=self.plot_settings)
 
         # Check rsun mismatch
-        rsun_map = self.meta["rsun_ref"]
-        rsun_target = outmap.meta["rsun_ref"]
-        if rsun_map != rsun_target:
+        if self.rsun_meters != outmap.rsun_meters:
             warn_user("rsun mismatch detected: "
-                      f"{self.name}.rsun_obs={self.rsun_obs}; {outmap.name}.rsun_obs={outmap.rsun_obs}. "
+                      f"{self.name}.rsun_meters={self.rsun_meters}; {outmap.name}.rsun_meters={outmap.rsun_meters}. "
                       "This might cause unexpected results during reprojection.")
 
         if return_footprint:
