### 2024-01-24 #147 @nabobalis
I have written up https://github.com/sunpy/sunpy.org/pull/401 I am hoping to get some reviews on this post this week so it can be merged.
