### 2024-01-25 dbf7a49
moved diff_rot() to sunpy.sun.models
