### 2024-01-25 4910c4a
moved diff_rot() to sunpy.sun.models
