### 2024-01-25 1bdd674
moved diff_rot() to sunpy.sun.models
