+++ sunpy/coordinates/tests/test_transformations.py
@@ -40,7 +40,7 @@ from sunpy.coordinates import (
 )
 from sunpy.coordinates.ephemeris import get_body_heliographic_stonyhurst, get_earth
 from sunpy.coordinates.frames import _J2000
-from sunpy.physics.differential_rotation import diff_rot
+from sunpy.sun.models import diff_rot
 from sunpy.sun.constants import radius as _RSUN
 from sunpy.sun.constants import sidereal_rotation_rate
 from sunpy.time import parse_time
