### 2024-01-25 fd98ec3
Readme fixes and sdist verify check for cron
