+++ README.rst
@@ -1,108 +1,52 @@
 *********
 ``sunpy``
 *********
+|SunPy Logo|
 
-.. raw:: html
++-----------------------------------+-----------------------------------+-----------------------------------+
+|           Release                 |           Development             |           Community               |
++===================================+===================================+===================================+
+|       |Latest PyPi Version|       |        |Python Versions|          |         |Matrix Chat Room|        |
++-----------------------------------+-----------------------------------+-----------------------------------+
+|     |Latest Conda Version|        |     |Project Status: Active|      |                                   |
++-----------------------------------+-----------------------------------+-----------------------------------+
+|      |Zenodo - Latest DOI|        |  |Continuous Integration Status|  |    |Google Groups Mailing List|   |
++-----------------------------------+-----------------------------------+-----------------------------------+
+|    |sunpy stable documentation|   |     |CodeCov Code Coverage|       |       |Powered by NumFOCUS|       |
++-----------------------------------+-----------------------------------+-----------------------------------+
+|                                   |                                   |            |pyOpenSci|            |
++-----------------------------------+-----------------------------------+-----------------------------------+
 
-    <img
-    src="https://raw.githubusercontent.com/sunpy/sunpy-logo/master/sunpy_logo_landscape.svg"
-    alt="SunPy Logo"
-    align="center"
-    width="200px" />
-    <table>
-    <thead>
-        <tr>
-        <th> Release </th>
-        <th> Development </th>
-        <th> Community </th>
-        </tr>
-    </thead>
-    <tbody>
-        <tr>
-        <td>
-            <a href="https://pypi.python.org/pypi/sunpy/"
-            ><img src="https://img.shields.io/pypi/v/sunpy.svg" alt="Latest PyPi Version"
-            /></a>
-        </td>
-        <td>
-            <a href="https://pypi.python.org/pypi/sunpy/"
-            ><img src="https://img.shields.io/pypi/pyversions/sunpy" alt="Python Versions"
-            /></a>
-        </td>
-        <td>
-            <a href="https://app.element.io/#/room/#sunpy:openastronomy.org"
-            ><img
-                src="https://img.shields.io/matrix/sunpy:openastronomy.org.svg?colorB=%23FE7900&label=Chat&logo=matrix&server_fqdn=matrix.org"
-                alt="Matrix Chat Room"
-            /></a>
-        </td>
-        </tr>
-        <tr>
-        <td>
-            <a href="https://anaconda.org/conda-forge/sunpy"
-            ><img src="https://anaconda.org/conda-forge/sunpy/badges/version.svg" alt="Latest Conda Version"
-            /></a>
-        </td>
-        <td>
-            <a href="https://www.repostatus.org/#active"
-            ><img
-                src="https://www.repostatus.org/badges/latest/active.svg"
-                alt="Project Status: Active - The project has reached a stable, usable state and is being actively developed."
-            /></a>
-        </td>
-        <td>
-            <a href="https://community.openastronomy.org/"
-            ><img
-                src="https://cdck-file-uploads-global.s3.dualstack.us-west-2.amazonaws.com/try2/original/1X/5e1e3b3cada2d7fbae4734d4bc53999933d71c95.svg"
-                height="20px"
-                alt="OpenAstronomy Discourse community"
[…]
