### 2024-01-25 f93627c
removed unnecessary lines and added change log
