+++ sunpy/sun/tests/test_models.py
@@ -10,36 +10,6 @@ from sunpy.sun.models import diff_rot
 
 
 @pytest.fixture
-def all_off_disk_map(aia171_test_map):
-    return aia171_test_map.submap((1, 1)*u.pix, top_right=(11, 12)*u.pix)
-
-
-@pytest.fixture
-def all_on_disk_map(aia171_test_map):
-    return aia171_test_map.submap((30, 60)*u.pix, top_right=(50, 85)*u.pix)
-
-
-@pytest.fixture
-def straddles_limb_map(aia171_test_map):
-    return aia171_test_map.submap((64, 80)*u.pix, top_right=(120, 127)*u.pix)
-
-
-@pytest.fixture
-def aia171_test_map_with_mask(aia171_test_map):
-    shape = aia171_test_map.data.shape
-    mask = np.zeros_like(aia171_test_map.data, dtype=bool)
-    mask[0:shape[0]//2, 0:shape[1]//2] = True
-    return sunpy.map.Map(np.ma.array(aia171_test_map.data, mask=mask), aia171_test_map.meta)
-
-
-@pytest.fixture
-def aia171_test_submap(aia171_test_map):
-    bl = SkyCoord(-512 * u.arcsec, 100 * u.arcsec, frame=aia171_test_map.coordinate_frame)
-    ur = SkyCoord(-100 * u.arcsec, 400 * u.arcsec, frame=aia171_test_map.coordinate_frame)
-    return aia171_test_map.submap(bl, top_right=ur)
-
-
-@pytest.fixture
 def seconds_per_day():
     return 24 * 60 * 60.0 * u.s
 
