### 2024-01-25 7ea23dd
removed unnecessary lines and added change log
