+++ setup.cfg
@@ -27,7 +27,6 @@ classifiers =
   Operating System :: OS Independent
   Programming Language :: Python
   Programming Language :: Python :: 3
-  Programming Language :: Python :: 3.9
   Programming Language :: Python :: 3.10
   Programming Language :: Python :: 3.11
   Programming Language :: Python :: 3.12
@@ -35,7 +34,7 @@ classifiers =
 
 [options]
 zip_safe = False
-python_requires = >=3.9
+python_requires = >=3.10
 packages = find:
 include_package_data = True
 setup_requires =
@@ -148,7 +147,7 @@ asdf.extensions =
 
 [bdist_wheel]
 # the Py_LIMITED_API version hex in _pyana.c should match the version specified here
-py_limited_api = cp39
+py_limited_api = cp310
 
 [extension-helpers]
 use_extension_helpers = true
