+++ sunpy/sun/tests/test_models.py
@@ -2,10 +2,9 @@ import numpy as np
 import pytest
 
 import astropy.units as u
-from astropy.coordinates import Longitude, SkyCoord
+from astropy.coordinates import Longitude
 from astropy.tests.helper import assert_quantity_allclose
 
-import sunpy.map
 from sunpy.sun.models import diff_rot
 
 
