### 2024-01-25 b655379
Backport PR #7262: Fix map metadata how to
