### 2024-01-25 bff75a9
Update docs/whatsnew/6.0.rst
Co-authored-by: David Stansby <dstansby@gmail.com>
