### 2024-01-25 f97755b
changelog for the opencv group
