+++ changelog/7383.trivial.1.rst
@@ -1,0 +1,5 @@
+Added a new optional extra group to install "opencv" if you want to it for affine transform.
+
+.. code-block:: bash
+
+    pip install sunpy[opencv]
