### 2024-01-25 8245187
Min pins for docs & test groups
