+++ setup.cfg
@@ -43,7 +43,7 @@ setup_requires =
 install_requires =
   astropy>=5.2.0
   numpy>=1.21.0
-  packaging>=19.0
+  packaging>=21.3
   parfive[ftp]>=2.0.0
   requests
 
@@ -99,38 +99,35 @@ all =
   %(visualization)s
 tests =
   %(all)s
-  hvpy
-  jplephem  # For some coordinates tests
+  hvpy>=1.1.0
+  jplephem>=2.14  # For some coordinates tests
   %(opencv)s
-  pytest-astropy
-  pytest-doctestplus
-  pytest-mock
-  pytest-mpl
-  pytest-xdist
-  pytest >= 7.1.0
+  pytest-astropy>=0.11.0
+  pytest-mpl>=0.16
+  pytest-xdist>=3.0.2
+  pytest>=7.1.0
   %(spice)s
 docs =
   %(all)s
-  hvpy
-  packaging
-  ruamel.yaml
-  sphinx
-  sphinx-automodapi
+  hvpy>=1.1.0
+  ruamel.yaml>=0.17.19
+  sphinx>=5.0.0
+  sphinx-automodapi>=0.14.1
   # Remove next line when fixed in sphinx-changelog; see https://github.com/OpenAstronomy/sphinx-changelog/issues/22
   importlib-resources<6
-  sphinx-changelog
-  sphinx-copybutton
-  sphinx-design
-  sphinx-gallery
-  sphinxext-opengraph
+  sphinx-changelog>=1.2.0
+  sphinx-copybutton>=0.5.0
+  sphinx-design>=0.2.0
+  sphinx-gallery>=0.11.0
+  sphinxext-opengraph>=0.6.0
   %(spice)s  # Importing sunpy.coordinates.spice raises an error if spiceypy is not installed
-  sunpy-sphinx-theme
-  sphinx-hoverxref
+  sunpy-sphinx-theme>=2.0.3
+  sphinx-hoverxref>=1.3.0
 docs-gallery =
   %(docs)s
-  astroquery
-  jplephem
-  mplcairo
+  astroquery>=0.4.6
+  jplephem>=2.14
+  mplcairo>=0.5
   %(opencv)s
 dev =
   %(all)s
