### 2024-01-25 3eaac17
Update changelog/7383.trivial.rst
Co-authored-by: David Stansby <dstansby@gmail.com>
