### 2024-01-25 be1ac50
Merge pull request #7397 from nabobalis/devdeps
Add pyerfa as core dep and fix imports (fixes devdeps)

(cherry picked from commit 3f9d063adae5bd36be6df2a0712b47b829a26119)
