+++ setup.cfg
@@ -46,6 +46,8 @@ install_requires =
   numpy>=1.21.0
   packaging>=19.0
   parfive[ftp]>=2.0.0
+  pyerfa
+  requests
 
 [options.extras_require]
 # The list of available extras is also in the installation docs, if you add or remove one please edit it there as well.
