### 2024-01-25 b873b11
Update setup.cfg
