+++ setup.cfg
@@ -47,7 +47,6 @@ install_requires =
   packaging>=19.0
   parfive[ftp]>=2.0.0
   pyerfa
-  requests
 
 [options.extras_require]
 # The list of available extras is also in the installation docs, if you add or remove one please edit it there as well.
