### 2024-01-26 1b2eb70
Merge pull request #7400 from nabobalis/readme
Readme fixes and sdist verify check for cron

(cherry picked from commit 74eae8e02900a7911646f7fbb39fcb14054b942e)
