+++ README.rst
@@ -1,30 +1,21 @@
 *********
 ``sunpy``
 *********
-
 |SunPy Logo|
 
-+----------------------+----------------------+----------------------+
-| Release              | Development          | Community            |
-+======================+======================+======================+
-| |Latest PyPi         | |Python Versions|    | |Matrix Chat Room|   |
-| Version|             |                      |                      |
-+----------------------+----------------------+----------------------+
-| |Latest Conda        | |Project Status:     | |OpenAstronomy       |
-| Version|             | Active - The project | Discourse community| |
-|                      | has reached a        |                      |
-|                      | stable, usable state |                      |
-|                      | and is being         |                      |
-|                      | actively developed.| |                      |
-+----------------------+----------------------+----------------------+
-| |Zenodo - Latest     | |Continuous          | |Google Groups       |
-| DOI|                 | Integration Status|  | Mailing List|        |
-+----------------------+----------------------+----------------------+
-| |sunpy stable        | |CodeCov Code        | |Powered by          |
-| documentation|       | Coverage|            | NumFOCUS|            |
-+----------------------+----------------------+----------------------+
-|                      |                      | |pyOpenSci|          |
-+----------------------+----------------------+----------------------+
++-----------------------------------+-----------------------------------+-----------------------------------+
+|           Release                 |           Development             |           Community               |
++===================================+===================================+===================================+
+|       |Latest PyPi Version|       |        |Python Versions|          |         |Matrix Chat Room|        |
++-----------------------------------+-----------------------------------+-----------------------------------+
+|     |Latest Conda Version|        |     |Project Status: Active|      |     |OpenAstronomy Discourse|     |
++-----------------------------------+-----------------------------------+-----------------------------------+
+|      |Zenodo - Latest DOI|        |  |Continuous Integration Status|  |    |Google Groups Mailing List|   |
++-----------------------------------+-----------------------------------+-----------------------------------+
+|    |sunpy stable documentation|   |     |CodeCov Code Coverage|       |       |Powered by NumFOCUS|       |
++-----------------------------------+-----------------------------------+-----------------------------------+
+|         |sunpy citation|          |                                   |            |pyOpenSci|            |
++-----------------------------------+-----------------------------------+-----------------------------------+
 
 .. |SunPy Logo| image:: https://raw.githubusercontent.com/sunpy/sunpy-logo/master/generated/sunpy_logo_landscape.png
    :width: 200px
@@ -36,16 +27,16 @@
    :target: https://app.element.io/#/room/#sunpy:openastronomy.org
 .. |Latest Conda Version| image:: https://anaconda.org/conda-forge/sunpy/badges/version.svg
    :target: https://anaconda.org/conda-forge/sunpy
-.. |Project Status: Active - The project has reached a stable, usable state and is being actively developed.| image:: https://www.repostatus.org/badges/latest/active.svg
+.. |Project Status: Active| image:: https://www.repostatus.org/badges/latest/active.svg
    :target: https://www.repostatus.org/#active
-.. |OpenAstronomy Discourse community| image:: https://cdck-file-uploads-global.s3.dualstack.us-west-2.amazonaws.com/try2/original/1X/5e1e3b3cada2d7fbae4734d4bc53999933d71c95.svg
[…]
