### 2024-01-26 b1f278f
Update docs/whatsnew/6.0.rst
