### 2024-01-26 50c7572
Update docs/dev_guide/contents/newcomers.rst
Co-authored-by: Albert Y. Shih <ayshih@gmail.com>
