### 2024-01-26 096e971
Apply suggestions from code review
Co-authored-by: David Stansby <dstansby@gmail.com>
