### 2024-01-26 2a2f3cd
Update figure envs to be more modern
