+++ tox.ini
@@ -66,9 +66,9 @@ deps =
     oldestdeps: zeep<4.2.0
     # Figure tests need a tightly controlled environment
     # These are pinned to the minimum version that we support
-    figure-!devdeps: astropy==5.3.4
-    figure-!devdeps: matplotlib==3.5.3
-    figure-!devdeps: mpl-animators==1.0.0
+    figure-!devdeps: astropy==6.0.0
+    figure-!devdeps: matplotlib==3.8.2
+    figure-!devdeps: mpl-animators==1.1.1
 extras =
     all
     tests
