+++ sunpy/net/hek2vso/hek2vso.py
@@ -181,14 +181,15 @@ class H2VClient:
         >>> res = h2v.translate_and_query(q)  # doctest: +REMOTE_DATA
         """
         vso_query = translate_results_to_query(hek_results)
-
+        kwargs = {}
+        if vso_response_format != "table":
+            kwargs = {"response_format": vso_response_format}
         for query in tqdm(vso_query, unit="records"):
-            temp = self.vso_client.search(*query, response_format=vso_response_format)
+            temp = self.vso_client.search(*query, **kwargs)
             self.vso_results.append(temp)
             self.num_of_records += len(temp)
-            if limit is not None:
-                if self.num_of_records >= limit:
-                    break
+            if limit is not None and self.num_of_records >= limit:
+                break
 
         return self.vso_results
 
