### 2024-01-26 f699828
added matching to unit test. some other refactoring
