### 2024-01-26 9fe0501
better phrasing of change log
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
