### 2024-01-27 b4d9755
fix missing link to astropy
