### 2024-01-27 #147 @cmarmo
Thank you @nabobalis: without further ado let's declare sunPy accepted into the pyOpenSci ecosystem!
I'm going to close this issue and move the announcement on the pyOpenSci discourse.

Thank you to all people involved in the process!
